% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99 tools.R
\name{combine}
\alias{combine}
\title{Title}
\usage{
combine(A, B)
}
\arguments{
\item{B}{}
}
\description{
Title
}
\examples{
A <- matrix(c(1, 0.8, 0.8, 1), nrow = 2, ncol = 2, byrow = TRUE)
combine(A, A)

A <- c(1, 2, 3)
B <- c(4, 5, 6)
combine(A, B)
}
