% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conProb.R
\name{conProb}
\alias{conProb}
\title{Consistency probability for MRCT via criterion Method I of MHLW (2007) (conditional version) under fixed effects model}
\usage{
conProb(
  alpha,
  power,
  pi = 0.5,
  rF,
  d,
  sigmaTrt,
  sigmaCtrl = sigmaTrt,
  randRatio = 1
)
}
\arguments{
\item{alpha}{The Type I error.}

\item{power}{Power.}

\item{pi}{The threshold ratio in criterion Method I of MHLW (2007) (conditional version). Defaults to 0.5.}

\item{rF}{The regional fraction.}

\item{d}{The true mean of difference of response.}

\item{sigmaTrt}{The standard deviation of response in the treatment group.}

\item{sigmaCtrl}{The standard deviation of response in the control group. Defaults to \code{sigmaTrt}.}

\item{randRatio}{The randomization ratio between the treatment group and control group. Defaults to 1.}
}
\value{
A list containing the following two components:
\describe{
\item{\code{CP}}{The consistency probability, a scalar.}
\item{\code{N}}{The overall sample size.}
}
}
\description{
Calculate the consistency probability for MRCT via criterion Method I of MHLW (2007) (conditional version) under fixed effects model.
}
\details{
The consistency probability via criterion Method I of MHLW (2007) (conditional version),
\deqn{\mathrm{Pr}\left(D_{k}\ge \pi D\ |\ T>z_{1-\alpha}\right),}
is approximately
\deqn{
 \frac{1}{1-\beta}\int_{-z_{1-\beta}}^{\infty} \Phi\left(\frac{(1-\pi)(u + z_{1-\alpha}+z_{1-\beta})}{\sqrt{f_{k}^{-1}-1}}\right)\phi(u)du.
}
Since there is no closed form of above equation, \code{conProb} utilizes the \code{\link[stats]{integrate}} function for numerical integration.

As the first equation includes none of \code{d}, \code{sigmaTrt},
\code{sigmaCtrl} and \code{randRatio}, if only the consistency probability
is considered, then the values of \code{d} and \code{sigmaTrt} could be arbitrary.

The overall sample size is calculated based on the following equation,
\deqn{
N^{(\textrm{c})}=\frac{\left\{r^{-1}\sigma^{2(\textrm{t})}+\sigma^{2(\textrm{c})}\right\}(z_{1-\alpha}+z_{1-\beta})^{2}}{d^{2}}, \quad N^{(\textrm{t})}=rN^{(\textrm{c})}.
}
Then \eqn{N = N^{(\textrm{t})} + N^{(\textrm{c})}}.
}
\examples{

conProb(alpha = 0.05, power = 0.8, rF = 0.271, d = 1, sigmaTrt = 4)

}
