% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conProb2MRCTs.R
\name{conProb2MRCTs}
\alias{conProb2MRCTs}
\title{Consistency probability for two MRCTs in general}
\usage{
conProb2MRCTs(
  alpha,
  power1,
  power2,
  pi,
  rF1,
  rF2,
  d1,
  d2,
  sigmaTrt1,
  sigmaCtrl1,
  sigmaTrt2,
  sigmaCtrl2,
  randRatio1,
  randRatio2
)
}
\arguments{
\item{alpha}{The type I error.}

\item{power1}{Power for MRCT 1.}

\item{power2}{Power for MRCT 2.}

\item{pi}{The threshold ratio in Japan's criterion I (conditional version).}

\item{rF1}{The regional fraction for MRCT 1.}

\item{rF2}{The regional fraction for MRCT 2.}

\item{d1}{The true mean of difference of response for MRCT 1.}

\item{d2}{The true mean of difference of response for MRCT 2.}

\item{sigmaTrt1}{The standard deviation of response in the treatment group for MRCT 1.}

\item{sigmaCtrl1}{The standard deviation of response in the control group for MRCT 1.}

\item{sigmaTrt2}{The standard deviation of response in the treatment group for MRCT 2.}

\item{sigmaCtrl2}{The standard deviation of response in the control group for MRCT 2.}

\item{randRatio1}{The randomization ratio between the treatment group and control group for MRCT 1.}

\item{randRatio2}{The randomization ratio between the treatment group and control group for MRCT 2.}
}
\value{
CP The consistency probability, a scalar.
}
\description{
Calculate the (conditional) consistency probability for two MRCTs in general
}
\details{
The extended (conditional) consistency probability, \eqn{\mathrm{Pr}\left(D_{k,\textrm{pool}}\ge \pi D_\textrm{pool}\ |\ T^{(1)}>z_{1-\alpha},T^{(2)}>z_{1-\alpha}\right)},
equals
\deqn{
 \begin{aligned}
 &\frac{1}{(1-\beta_{1})(1-\beta_{2})}\int_{-z_{1-\beta_{1}}}^{\infty}\int_{-z_{1-\beta_{2}}}^{\infty} \\
 &\quad
 \Phi\left(\frac{(1-\pi)\left\{w^{(1)}\sigma^{(1)}_{d}u+w^{(2)}\sigma^{(2)}_{d}v + w^{(1)}d^{(1)} + w^{(2)}d^{(2)}\right\}}{\sqrt{\left\{\left(f_{k}^{(1)}\right)^{-1}-1\right\} \left(w^{(1)}\sigma^{(1)}_{d}\right)^{2} + \left\{\left(f_{k}^{(2)}\right)^{-1}-1\right\} \left(w^{(2)}\sigma^{(2)}_{d}\right)^{2}}} \right)\phi(u)\phi(v)dudv.
 \end{aligned}
}
Since there is no closed form of above equation, \code{conProb2MRCTs} utilizes
the \code{\link[cubature]{adaptIntegrate}} function for numerical integration.
}
\examples{
\dontrun{
conProb2MRCTs(0.025,0.8,0.8,0.5,0.127,0.127,1,1,4,4,4,4,1,1)
}


}
