% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conProb2MRCTsH.R
\name{conProb2MRCTsH}
\alias{conProb2MRCTsH}
\title{Consistency probability for two homogenous MRCTs}
\usage{
conProb2MRCTsH(alpha, power1, power2, pi, rF1, rF2)
}
\arguments{
\item{alpha}{The type I error.}

\item{power1}{Power for MRCT 1.}

\item{power2}{Power for MRCT 2.}

\item{pi}{The threshold ratio in Japan's criterion I (conditional version).}

\item{rF1}{The regional fraction for MRCT 1.}

\item{rF2}{The regional fraction for MRCT 2.}
}
\value{
CP The consistency probability, a scalar.
}
\description{
Calculate the (conditional) consistency probability for two homogeneous MRCTs
}
\details{
The extended (conditional) consistency probability, \eqn{\mathrm{Pr}\left(D_{k,\textrm{pool}}\ge \pi D_\textrm{pool}\ |\ T^{(1)}>z_{1-\alpha},T^{(2)}>z_{1-\alpha}\right)},
equals
\deqn{
 \begin{aligned}
 &\frac{1}{(1-\beta_{1})(1-\beta_{2})}\int_{-(z_{1-\beta_{1}}+z_{1-\beta_{2}})/\sqrt{2}}^{\infty}\left[ \Phi\left(\frac{(1-\pi)\left\{\sqrt{2}u + (2z_{1-\alpha}+z_{1-\beta_{1}}+z_{1-\beta_{2}})\right\}}{\sqrt{\left(f_{k}^{(1)}\right)^{-1}+\left(f_{k}^{(2)}\right)^{-1}-2}}\right) \right. \\
 &\qquad\qquad\qquad\qquad \left.\left\{ \Phi\left(u+\sqrt{2}z_{1-\beta_{1}}\right)+\Phi\left(u+\sqrt{2}z_{1-\beta_{2}}\right)-1\right\} \right] \phi(u)du,
 \end{aligned}
}
when assuming homogeneous variances, equal treatment effects
and equal randomization ratios across two studies,
Since there is no closed form of above equation, \code{conProb2MRCTsH} utilizes
the \code{\link[stats]{integrate}} function for numerical integration.
}
\examples{
\dontrun{
conProb2MRCTsH(0.025,0.8,0.8,0.5,0.127,0.127)
}

}
