% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conProbII.R
\name{conProbII}
\alias{conProbII}
\title{Consistency probability for one MRCT via Japan's criterion II (conditional version)}
\usage{
conProbII(
  alpha,
  power,
  rF,
  d,
  sigmaTrt,
  sigmaCtrl = sigmaTrt,
  pTrt,
  pCtrl,
  randRatio = 1,
  responseType = c("continuous", "binary"),
  B = 1e+05
)
}
\arguments{
\item{alpha}{The Type I error.}

\item{power}{Power.}

\item{rF}{The regional fractions, a \code{K x 1} vector with each component
representing the regional fraction in region \eqn{k} and the sum must equal 1.}

\item{d}{The true mean of difference of response.}

\item{sigmaTrt}{The standard deviation of response in the treatment group.}

\item{sigmaCtrl}{The standard deviation of response in the control group. Defaults to \code{sigmaTrt}.}

\item{pTrt}{The mean of the response in the treatment group.}

\item{pCtrl}{The mean of the response in the control group.}

\item{randRatio}{The randomization ratio between the treatment group and control group. Defaults to 1.}

\item{responseType}{The type of response. One of "continuous" and "binary".}

\item{B}{The number of simulation by Monto Carlo for \code{responseType} = "binary". Defaults to 100,000.}
}
\value{
A list containing the following two components:
\describe{
\item{\code{CP}}{The consistency probability, a scalar.}
\item{\code{N}}{The overall sample size.}
}
}
\description{
Calculate the consistency probability for one MRCT via Japan's criterion II (conditional version).
}
\details{
The consistency probability via Japan's criterion II (conditional version),
\deqn{\mathrm{Pr}\left(D_{k}\geq 0, k=1,\ldots,K\ |\ T>z_{1-\alpha}\right),}
is approximately
\deqn{
 \frac{1}{1-\beta}\int_{-z_{1-\beta}}^{\infty} \prod_{k=1}^{K}\Phi\left(\frac{u + z_{1-\alpha}+z_{1-\beta}}{\sqrt{f_{k}^{-1}-1}}\right)\phi(u)du.
}
Since there is no closed form of above equation, \code{conProbII} utilizes the \code{\link[stats]{integrate}} function for numerical integration.

As the first equation includes none of \code{d}, \code{sigmaTrt},
\code{sigmaCtrl} and \code{randRatio}, if only the consistency probability
is considered, then the values of \code{d} and \code{sigmaTrt} could be arbitrary.

For binary response, the above approximation loses precision under moderate sample size.
Hence \code{conProbII} applies Monto Carlo to calculate the correct consistency probability.

The overall sample size is calculated in the same way as \code{\link{conProb}}.
}
\examples{
### Example 1
alpha <- 0.05
power <- 0.8 
conProbII(alpha, power, rF = rep(1/2,2), d = 1, sigmaTrt = 4, responseType = "continuous")
conProbII(alpha, power, rF = rep(1/3,3), d = 1, sigmaTrt = 4, responseType = "continuous")
conProbII(alpha, power, rF = rep(1/4,4), d = 1, sigmaTrt = 4, responseType = "continuous") 
rFk1 <- 0.101
rFk2 <- rFk3 <- (1-rFk1)/2
rF <- c(rFk1, rFk2, rFk3)
conProbII(alpha, power, rF = rF, d = 1, sigmaTrt = 4, responseType = "continuous") 

### Example 2
rFk1 <- 0.149
rFk2 <- rFk3 <- (1-rFk1)/2
rF <- c(rFk1, rFk2, rFk3)
set.seed(123)
conProbII(alpha, power, rF = rF, pTrt = 0.8, pCtrl = 0.7, responseType = "binary")
rFk1 <- 0.14
rFk2 <- rFk3 <- (1-rFk1)/2
rF <- c(rFk1, rFk2, rFk3)
set.seed(123)
conProbII(alpha, power, rF = rF, pTrt = 0.7, pCtrl = 0.6, responseType = "binary") 
rFk1 <- 0.101
rFk2 <- rFk3 <- (1-rFk1)/2
rF <- c(rFk1, rFk2, rFk3)
set.seed(123)
conProbII(alpha, power, rF = rF, pTrt = 0.8, pCtrl = 0.7, responseType = "binary") 
conProbII(alpha, power, rF = rF, pTrt = 0.7, pCtrl = 0.6, responseType = "binary") 

}
