% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conProbII2.R
\name{conProbII2}
\alias{conProbII2}
\title{Consistency probability for two MRCTs via Japan's criterion II (conditional version)}
\usage{
conProbII2(
  alpha,
  power1,
  power2 = power1,
  rF1,
  rF2 = rF1,
  d1,
  d2 = d1,
  sigmaTrt1,
  sigmaCtrl1 = sigmaTrt1,
  sigmaTrt2 = sigmaTrt1,
  sigmaCtrl2 = sigmaTrt2,
  pTrt1,
  pCtrl1,
  pTrt2 = pTrt1,
  pCtrl2 = pCtrl1,
  randRatio1 = 1,
  randRatio2 = randRatio1,
  responseType = c("continuous", "binary"),
  B = 1e+05
)
}
\arguments{
\item{alpha}{The Type I error.}

\item{power1}{Power for MRCT 1.}

\item{power2}{Power for MRCT 2. Defaults to \code{power1}.}

\item{rF1}{The regional fractions for MRCT 1, a \code{K x 1} vector with each component
representing the regional fraction in region \eqn{k} and the sum must equal 1.}

\item{rF2}{The regional fractions for MRCT 2, a \code{K x 1} vector with each component
representing the regional fraction in region \eqn{k} and the sum must equal 1. Defaults to \code{rF1}.}

\item{d1}{The true mean of difference of response for MRCT 1.}

\item{d2}{The true mean of difference of response for MRCT 2. Defaults to \code{d1}.}

\item{sigmaTrt1}{The standard deviation of response in the treatment group for MRCT 1.}

\item{sigmaCtrl1}{The standard deviation of response in the control group for MRCT 1. Defaults to \code{sigmaTrt1}.}

\item{sigmaTrt2}{The standard deviation of response in the treatment group for MRCT 2. Defaults to \code{sigmaTrt1}.}

\item{sigmaCtrl2}{The standard deviation of response in the control group for MRCT 2. Defaults to \code{sigmaTrt2}.}

\item{pTrt1}{The mean of the response in the treatment group for MRCT 1.}

\item{pCtrl1}{The mean of the response in the control group for MRCT 1.}

\item{pTrt2}{The mean of the response in the treatment group for MRCT 2. Defaults to \code{pTrt1}.}

\item{pCtrl2}{The mean of the response in the control group for MRCT 2. Defaults to \code{pTrt2}.}

\item{randRatio1}{The randomization ratio between the treatment group and control group for MRCT 1.}

\item{randRatio2}{The randomization ratio between the treatment group and control group for MRCT 2. Defaults to \code{randRatio1}.}

\item{responseType}{The type of response. One of "continuous" and "binary".}

\item{B}{The number of simulation by Monto Carlo for \code{responseType} = "binary". Defaults to 100,000.}
}
\value{
A list containing the following two components:
\describe{
\item{\code{CP}}{The consistency probability, a scalar.}
\item{\code{N1}}{The overall sample size for MRCT 1.}
\item{\code{N2}}{The overall sample size for MRCT 2.}
}
}
\description{
Calculate the consistency probability for two MRCTs via Japan's criterion II (conditional version).
}
\details{
The extended consistency probability via the extended Japan's criterion II (conditional version),
\deqn{\mathrm{Pr}\left(D_{k,\textrm{pool}}\geq 0, k=1,\cdots,K\ |\ T^{(1)}>z_{1-\alpha},T^{(2)}>z_{1-\alpha}\right),}
is approximately
\deqn{
 \begin{aligned} 
   &\frac{1}{(1-\beta_{1})(1-\beta_{2})}\int_{-z_{1-\beta_{1}}}^{\infty}\int_{-z_{1-\beta_{2}}}^{\infty}  \notag\\ 
   &\quad\prod_{k=1}^{K}\Phi\left(\frac{ w^{(1)}\sigma^{(1)}_{d}u+w^{(2)}\sigma^{(2)}_{d}v + w^{(1)}d^{(1)} + w^{(2)}d^{(2)} }{\sqrt{\left\{\left(f_{k}^{(1)}\right)^{-1}-1\right\} \left(w^{(1)}\sigma^{(1)}_{d}\right)^{2} + \left\{\left(f_{k}^{(2)}\right)^{-1}-1\right\} \left(w^{(2)}\sigma^{(2)}_{d}\right)^{2}}} \right)\phi(u)\phi(v)dudv,
\end{aligned} 
}
where \eqn{w^{(s)}} and \eqn{\sigma_d^{2}} are defined in \code{\link{conProb2}}.

Since there is no closed forms of above equations, \code{conProbII2} utilizes
the \code{\link[cubature]{adaptIntegrate}} function for numerical integration.

For binary response, the above approximation loses precision under moderate sample size.
Hence \code{conProbII2} applies Monto Carlo to calculate the correct consistency probability.

The overall sample size is calculated in the same way as \code{\link{conProb}}.
But additionally requiring all of \eqn{N^{(\textrm{t},s)}_{k}} and \eqn{N^{(\textrm{c},s)}_{k}}
should be integers and hence \eqn{N^{(\textrm{t},s)}}, \eqn{N^{(\textrm{c,s})}} and \eqn{N^{s}}.
}
\examples{

### Example 3
alpha <- 0.05
power1 <- 0.8 
conProbII2(alpha, power1, rF1 = rep(1/2,2), d1 = 1, sigmaTrt1 = 4, responseType = "continuous")
conProbII2(alpha, power1, rF1 = rep(1/3,3), d1 = 1, sigmaTrt1 = 4, responseType = "continuous")
conProbII2(alpha, power1, rF1 = rep(1/4,4), d1 = 1, sigmaTrt1 = 4, responseType = "continuous") 
rF11 <- 0.044
rF12 <- rF13 <- (1-rF11)/2
rF1 <- c(rF11, rF12, rF13)
conProbII2(alpha, power1, rF1 = rF1, d1 = 1, sigmaTrt1 = 4, responseType = "continuous") 

### Example 4
rF11 <- 0.06
rF12 <- rF13 <- (1-rF11)/2
rF1 <- c(rF11, rF12, rF13)
set.seed(123)
conProbII2(alpha, power1, rF1 = rF1, pTrt1 = 0.8, pCtrl1 = 0.7, responseType = "binary") 
rF11 <- 0.044
rF12 <- rF13 <- (1-rF11)/2
rF1 <- c(rF11, rF12, rF13)
set.seed(123)
conProbII2(alpha, power1, rF1 = rF1, pTrt1 = 0.8, pCtrl1 = 0.7, responseType = "binary") 

}
