% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPlbd.R
\name{cplbd}
\alias{cplbd}
\alias{CPlbd}
\title{Lower bound of the consistency probability for one MRCT via Japan's criterion I (conditional version) under random effects model.}
\usage{
CPlbd(alpha, beta, tau_delta_r, delta, sigma0r, sigma1r, fr, pi, k)
}
\arguments{
\item{alpha}{The one-sided Type I error.}

\item{beta}{The Type II error.}

\item{tau_delta_r}{The ratio of the standard deviation vs mean of the Gaussian prior of the regional treatment effect.}

\item{delta}{The mean of the Gaussian prior of the regional treatment effect.}

\item{sigma0r}{The standard deviation of response in the treatment group. Specified for region(s) of interest, either a scalar or vector of length equals to the total number of region(s) of interest.}

\item{sigma1r}{The standard deviation of response in the control group. Specified for all region(s), either a scalar or vector of length equals to the total number of region(s) of interest.}

\item{fr}{The regional fraction for the region(s) of interest, either a scalar or vector.}

\item{pi}{The threshold ratio in Japan's criterion I (conditional version). Defaults to 0.5.}

\item{k}{The randomization ratio between the treatment group and control group. Defaults to 1.}
}
\value{
A list containing the following components:
\describe{
\item{\code{CPlbd}}{The lower bound of consistency probability, a scalar.}
\item{\code{n0r}}{The regional sample size of the control group. (n.a. is returned when the condition for positive sample size is not met)}
\item{\code{n1r}}{The regional sample size of the treatment group. (n.a. is returned when the condition for positive sample size is not met)}
\item{\code{nr}}{The regional sample size.(n.a. is returned when the condition for positive sample size is not met)}

}
}
\description{
Calculate the lower bound of consistency probability for one MRCT via Japan's criterion I (conditional version) under random effects model for normal endpoint.
}
\details{
The lower bound of the consistency probability via Japan's criterion I (conditional version) under random effects model,
\eqn{ \mathrm{Pr} \left( \widetilde{D}_{r} \ge \pi \widetilde{D} | T > z_{1-\alpha} \right) },
is approximately
\deqn{
 \frac{1}{1-\beta}\int_{u>z_{\beta}}\Phi\left(\frac{2\delta^2(1-\pi)(u+z_{1-\alpha}+z_{1-\beta})}{\tau^2(z_{1-\alpha}+z_{1-\beta})^2} \right) d\Phi(u).
}
provided \eqn{\frac{\tau}{\delta}<\frac{\sqrt{2}}{z_{1-\alpha}+z_{1-\beta}}}
Since there is no closed form of above equation, \code{CPlbd} utilizes the \code{\link[stats]{integrate}} function for numerical integration.

The sample size for the region of interest corresponding to the lower bound is calculated based on the below equation,
\deqn{
\frac{h_r}{h_r+1}=\frac{\tau^2}{2\delta^2}(z_{1-\alpha}+z_{1-\beta})^2, \quad n_1^{(r)}=kn_0^{(r)}
}
where \eqn{h_r=\tau^2/\sigma^{2(r)}, \quad \sigma^{2(r)}=\Omega_r/(n_0f_r), \quad \Omega_r=k^{-1}\sigma_1^{2(r)}+\sigma_1^{2(r)}}
Then \eqn{n^{(r)} = n_0^{(r)}+n_1^{(r)}}. Additionally, both of
\eqn{n_0^{(r)}} and \eqn{n_1^{(r)}} should be integers and hence \eqn{n^{(r)}}.
}
\examples{

CPlbd(alpha=0.05, beta=0.2, tau_delta_r=0.4, pi=0.5, delta=0.25, sigma0r=1,
       sigma1r=1, k=1, fr=c(0.1,0.5))

}
