% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/56 getN_Bin_Equi_JM1.R
\name{getN_Bin_Equi_JM1}
\alias{getN_Bin_Equi_JM1}
\title{Title}
\usage{
getN_Bin_Equi_JM1(p1_j, p0_j, p1_nj, p0_nj, pi, cut, beta1, N, r)
}
\arguments{
\item{r}{}
}
\description{
Title
}
\examples{
getN_Bin_Equi_JM1(
  p1_j = 0.4, p0_j = 0.5, p1_nj = 0.4, p0_nj = 0.4,
  pi = 0.5, cut = 0.3, beta1 = 0.2, N = seq(100, 400, 100), r = 1
)
}
