% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2_getN_Bin_Super_JM1.R
\name{getN_Bin_JM1}
\alias{getN_Bin_JM1}
\alias{getN_Bin_Super_JM1}
\alias{getN_Bin_Noninf_JM1}
\alias{getN_Bin_Equi_JM1}
\title{Regional sample size allocation using Japan's Method 1 for binary endpoints}
\usage{
getN_Bin_Super_JM1(
  p1_a,
  p0_a,
  p1_j,
  p0_j,
  pi = 0.5,
  alpha = NA,
  beta = NA,
  beta1 = 0.2,
  N = NA,
  r = 1,
  scale = "RD"
)

getN_Bin_Noninf_JM1(
  p1_a,
  p0_a,
  p1_j,
  p0_j,
  pi = 0.5,
  cut,
  alpha = NA,
  beta = NA,
  beta1 = 0.2,
  N = NA,
  r = 1,
  scale = "RD",
  direct = 1
)

getN_Bin_Equi_JM1(
  p1_a,
  p0_a,
  p1_j,
  p0_j,
  pi = 0.5,
  cut,
  alpha = NA,
  beta = NA,
  beta1 = 0.2,
  N = NA,
  r = 1,
  scale = "RD",
  maxN = 1e+06
)
}
\arguments{
\item{p1_a}{Rate of treatment group globally.}

\item{p0_a}{Rate of the control group globally.}

\item{p1_j}{Rate of treatment group in target region.}

\item{p0_j}{Rate of control group in target region.}

\item{pi}{Proportion of global efficacy to retain. The default value is 0.5, which means retaining half of the efficacy.}

\item{alpha}{One-sided type I error rate for global success, which is used to calculate global sample size only when \code{N} is \code{NA}. The default value is 0.025.}

\item{beta}{Type II error rate for global success, which is used to calculate global sample size only when \code{N} is missing.}

\item{beta1}{Type II error rate for efficacy consistency between target region and globally. The default value is 0.2.}

\item{N}{Global sample size. When \code{N} is \code{NA} and \code{alpha} and \code{beta} are not \code{NAs}, \code{N} will be calculated automatically.}

\item{r}{Ratio of the sample sizes of the treatment group to the control group. The default value is 1.}

\item{scale}{Optional values are "RD" for rate difference, "RR" for relative risk, and "OR" for odds ratio. Default value is "RD".}

\item{cut}{A positive value for non-inferiority or equivalence margin. For RD, the margin is on the original scale. For RR and OR, the margin is on the log scale. For example, if the non-inferiority margins for RD, RR, OR are 0.2, 0.6, and 1.3, then the \code{cut = 0.2}, \code{cut = -log(0.6)}, and \code{cut = log(1.3)}, respectively.}

\item{direct}{If \code{direct = 1}, larger values of RD/RR/OR are preferable. If \code{direct = -1}, smaller values of RD/RR/OR are preferable.}

\item{maxN}{Maximum possible sample size (\code{N}) in equivalence design. Default value is 1e6.}
}
\value{
A data frame where \code{f} is required proportion of sample size allocated to the target region, and \code{Nj} is required sample size for the target region, calculated as \code{Nj = N * f}.
}
\description{
Based on Japan's Method 1, given the global sample size and marginal probability (power) of efficacy consistency between target region and globally, calculate the required sample size allocated to the target region, in clinical trials using superiority, non-inferiority, and equivalence designs with binary endpoints.
}
\details{
The global success criterion and the efficacy consistency criterion between target region and globally could be found in \code{\link{getPwr_Bin_Super_JM1}}.
}
\examples{
getN_Bin_Super_JM1(
  p1_a = 0.75, p0_a = 0.5, p1_j = 0.7, p0_j = 0.5, pi = 0.5, beta1 = 0.2,
  N = 200, r = 1, scale = "RD"
)

getN_Bin_Noninf_JM1(
  p1_a = 0.5, p0_a = 0.5, p1_j = 0.6, p0_j = 0.5, pi = 0.5, cut = log(1.6),
  alpha = 0.025, beta = 0.2, beta1 = 0.2, N = NA, r = 1, scale = "RR",
  direct = -1
)
}
\references{
1. Quan H, Li M, Chen J, et al. Assessment of Consistency of Treatment Effects in Multiregional Clinical Trials. Drug Information J. 2010;44(5):617-632. doi:10.1177/009286151004400509

2. Liao JJZ, Yu Z, Li Y. Sample size allocation in multiregional equivalence studies. Pharm Stat. 2018;17(5):570-577. doi:10.1002/pst.1871
}
\seealso{
\code{\link{getPwr_Bin_Super_JM1}}, \code{\link{getN_Bin_Super}}.
}
