% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/35 getN_Bin_Noninf.R
\name{getN_Bin_Noninf}
\alias{getN_Bin_Noninf}
\title{Title}
\usage{
getN_Bin_Noninf(p1 = p1, p0 = p0, cut, alpha, beta, N, r, direct = 1)
}
\arguments{
\item{direct}{}
}
\description{
Title
}
\examples{
(v <- getN_Bin_Noninf(
  p1 = 0.4, p0 = 0.45, cut = 0.2, alpha = 0.025,
  beta = 0.2, N = NA, r = 1
))
getN_Bin_Noninf(
  p1 = 0.4, p0 = 0.45, cut = 0.2, alpha = 0.025,
  beta = NA, N = v$N, r = 1
)
}
