% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/34 getN_Bin_Super.R
\name{getN_Bin_Super}
\alias{getN_Bin_Super}
\title{Title}
\usage{
getN_Bin_Super(p1, p0, alpha, beta, N, r, direct = 1)
}
\arguments{
\item{direct}{}
}
\description{
Title
}
\examples{
(v <- getN_Bin_Super(
  p1 = 0.4, p0 = 0.2, alpha = 0.025,
  beta = 0.2, N = NA, r = 1
))
getN_Bin_Super(
  p1 = 0.4, p0 = 0.2, alpha = 0.025,
  beta = NA, N = v$N, r = 1
)
}
