% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c1_getN_Con_Super.R
\name{getN_Con}
\alias{getN_Con}
\alias{getN_Con_Super}
\alias{getN_Con_Noninf}
\alias{getN_Con_Equi}
\title{Sample size and power for continuous endpoints}
\usage{
getN_Con_Super(delta, sigma, alpha = 0.025, beta = NA, N = NA, r = 1)

getN_Con_Noninf(
  delta,
  sigma,
  cut,
  alpha = 0.025,
  beta = NA,
  N = NA,
  r = 1,
  direct = 1
)

getN_Con_Equi(
  delta,
  sigma,
  cut,
  alpha = 0.025,
  beta = NA,
  N = NA,
  r = 1,
  maxN = 1e+06
)
}
\arguments{
\item{delta}{Mean difference between treatment and control groups.}

\item{sigma}{Common standard deviation.}

\item{alpha}{One-sided type I error rate. The default value is 0.025.}

\item{beta}{Type II error rate. When \code{beta} is \code{NA} and \code{N} is not \code{NA}, the power will be returned.}

\item{N}{Sample size. When \code{N} is \code{NA} and \code{beta} is not \code{NA}, the sample size will be returned.}

\item{r}{Ratio of the sample sizes of the treatment group to the control group. The default value is 1.}

\item{cut}{A positive value for non-inferiority or equivalence margin.}

\item{direct}{\code{direct = 1} indicates that a larger mean difference is preferable, while \code{direct = -1} indicates that a smaller mean difference is preferable.}

\item{maxN}{Maximum possible sample size(\code{N}) in equivalence design. Default value is 1e6.}
}
\value{
A data frame containing input parameters and returned sample size or power.
}
\description{
Calculating sample size when given power or power when given sample size for continuous endpoints.
}
\details{
Taking the larger mean difference is preferable as an example. Sample size calculation is based on the following Z test, with the success criterion:

in superiority design:
\deqn{Z = \frac{\hat \delta}{\hat \sigma\sqrt{\frac{1}{N / (r + 1)} + \frac{1}{Nr / (r + 1)}}} > \Phi^{-1}(1 - \alpha)}

in non-inferiority design:
\deqn{Z = \frac{\hat \delta + \Delta}{\hat \sigma\sqrt{\frac{1}{N / (r + 1)} + \frac{1}{Nr / (r + 1)}}} > \Phi^{-1}(1 - \alpha)}

in equivalence design:
\deqn{Z_u = \frac{\hat \delta + \Delta}{\hat \sigma\sqrt{\frac{1}{N / (r + 1)} + \frac{1}{Nr / (r + 1)}}} > \Phi^{-1}(1 - \alpha)\text{ and } Z_l = \frac{\hat \delta - \Delta}{\hat \sigma\sqrt{\frac{1}{N / (r + 1)} + \frac{1}{Nr / (r + 1)}}} < \Phi^{-1}(\alpha)}

Where \eqn{\hat \delta} is the mean difference between treatment and control groups, and \eqn{\Delta} is the non-inferiority or equivalence margin.
}
\examples{
(v <- getN_Con_Super(delta = 1.5, sigma = 4, alpha = 0.025, beta = 0.2, N = NA, r = 1))
getN_Con_Super(delta = 1.5, sigma = 4, alpha = 0.025, beta = NA, N = v$N, r = 1)

(v <- getN_Con_Noninf(
  delta = 1, sigma = 4, cut = 2, alpha = 0.025, beta = 0.2, N = NA, r = 1, direct = -1
))
getN_Con_Noninf(
  delta = 1, sigma = 4, cut = 2, alpha = 0.025, beta = NA, N = v$N, r = 1, direct = -1
)
}
