% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b1_getN_Con_Super_JM1.R
\name{getN_Con_JM1}
\alias{getN_Con_JM1}
\alias{getN_Con_Super_JM1}
\alias{getN_Con_Noninf_JM1}
\alias{getN_Con_Equi_JM1}
\title{Regional sample size allocation using Japan's Method 1 for continuous endpoints}
\usage{
getN_Con_Super_JM1(
  delta_a,
  delta_j,
  sigma,
  pi = 0.5,
  alpha = NA,
  beta = NA,
  beta1 = 0.2,
  N = NA,
  r = 1
)

getN_Con_Noninf_JM1(
  delta_a,
  delta_j,
  sigma,
  pi = 0.5,
  cut,
  alpha = NA,
  beta = NA,
  beta1 = 0.2,
  N = NA,
  r = 1,
  direct = 1
)

getN_Con_Equi_JM1(
  delta_a,
  delta_j,
  sigma,
  pi = 0.5,
  cut,
  alpha = NA,
  beta = NA,
  beta1 = 0.2,
  N = NA,
  r = 1,
  maxN = 1e+06
)
}
\arguments{
\item{delta_a}{A vector. Mean difference between treatment and control groups globally.}

\item{delta_j}{A vector. Mean difference between treatment and control groups in target region.}

\item{sigma}{A vector. Common standard deviation.}

\item{pi}{A vector. Proportion of global efficacy to retain. Default value is 0.5, which means retaining half of the efficacy.}

\item{alpha}{A vector. One-sided type I error rate for global success, which is used to calculate global sample size only when \code{N} is \code{NA}. Default value is 0.025.}

\item{beta}{A vector. Type II error rate for global success, which is used to calculate global sample size only when \code{N} is \code{NA}.}

\item{beta1}{A vector. Type II error rate for efficacy consistency between target region and globally. Default value is 0.2.}

\item{N}{A vector. Global sample size. When \code{N} is \code{NA} and \code{alpha} and \code{beta} are not \code{NA}, \code{N} will be calculated automatically.}

\item{r}{A vector. Ratio of sample sizes of treatment group to control group. Default value is 1.}

\item{cut}{A vector. Positive value for non-inferiority or equivalence margin.}

\item{direct}{\code{direct = 1} indicates that a larger mean difference is preferable, while \code{direct = -1} indicates that a smaller mean difference is preferable.}

\item{maxN}{Maximum possible sample size (\code{N}) in equivalence design. Default value is 1e+06.}
}
\value{
A data frame where \code{f} is required proportion of sample size allocated to the target region, and \code{Nj} is the required sample size for the target region, calculated as \code{Nj = N * f}.
}
\description{
Based on Japan's Method 1, given the global sample size and marginal probability (power) of efficacy consistency between target region and globally, calculate the required sample size allocated to the target region, in clinical trials using superiority, non-inferiority, and  equivalence designs with continuous endpoints.
}
\details{
The global success criterion and the efficacy consistency criterion between target region and globally could be found in \code{\link{getPwr_Con_Super_JM1}}.
}
\examples{
getN_Con_Super_JM1(
  delta_a = 0.7, delta_j = 0.5, sigma = 1, pi = 0.5, beta1 = 0.2, N = 100,
  r = 1
)

# Global sample size will be calculated based on alpha and beta.
getN_Con_Noninf_JM1(
  delta_a = 0, delta_j = 0.5, sigma = 4, pi = 0.5, cut = 2, alpha = 0.025,
  beta = 0.2, beta1 = 0.2, N = NA, r = 1, direct = -1
)
}
\references{
1. Quan H, Li M, Chen J, et al. Assessment of Consistency of Treatment Effects in Multiregional Clinical Trials. Drug Information J. 2010;44(5):617-632. doi:10.1177/009286151004400509

2. Liao JJZ, Yu Z, Li Y. Sample size allocation in multiregional equivalence studies. Pharm Stat. 2018;17(5):570-577. doi:10.1002/pst.1871
}
\seealso{
\code{\link{getPwr_Con_Super_JM1}}, \code{\link{getN_Con_Super}}.
}
