% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32 getN_Con_Noninf.R
\name{getN_Con_Noninf}
\alias{getN_Con_Noninf}
\title{Title}
\usage{
getN_Con_Noninf(delta, sigma, cut, alpha, beta, N, r, direct = 1)
}
\arguments{
\item{direct}{}
}
\description{
Title
}
\examples{
(v <- getN_Con_Noninf(
  delta = 0, sigma = 2, cut = 1, alpha = 0.025,
  beta = 0.2, N = NA, r = 1
))
getN_Con_Noninf(
  delta = 0, sigma = 2, cut = 1, alpha = 0.025,
  beta = NA, N = v$N, r = 1
)
}
