% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31 getN_Con_Super.R
\name{getN_Con_Super}
\alias{getN_Con_Super}
\title{Title}
\usage{
getN_Con_Super(delta, sigma, alpha, beta, N, r, direct = 1)
}
\arguments{
\item{direct}{}
}
\description{
Title
}
\examples{
(v <- getN_Con_Super(
  delta = 1, sigma = 4, alpha = 0.025,
  beta = 0.2, N = NA, r = 1
))
getN_Con_Super(
  delta = 1, sigma = 4, alpha = 0.025,
  beta = NA, N = v$N, r = 1
)
}
