% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/51 getN_Con_Super_JM1.R
\name{getN_Con_Super_JM1}
\alias{getN_Con_Super_JM1}
\title{Title}
\usage{
getN_Con_Super_JM1(
  delta_j,
  delta_nj,
  sigma,
  pi,
  alpha,
  beta1,
  N,
  r,
  direct = 1
)
}
\arguments{
\item{direct}{}
}
\description{
Title
}
\examples{
getN_Con_Super_JM1(
  delta_j = 0.5, delta_nj = 0.7, sigma = 1,
  pi = 0.5, alpha = 0.025, beta1 = 0.2, N = seq(100, 400, 100), r = 1, direct = 1
)
}
