% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4_getN_Count_Super.R
\name{getN_Count}
\alias{getN_Count}
\alias{getN_Count_Super}
\alias{getN_Count_Noninf}
\alias{getN_Count_Equi}
\title{Sample size and power for count endpoints}
\usage{
getN_Count_Super(
  delta,
  lambda0,
  t,
  k = 0,
  alpha = 0.025,
  beta = NA,
  N = NA,
  r = 1
)

getN_Count_Noninf(
  delta,
  lambda0,
  t,
  k = 0,
  cut,
  alpha = 0.025,
  beta = NA,
  N = NA,
  r = 1,
  direct = 1
)

getN_Count_Equi(
  delta,
  lambda0,
  t,
  k = 0,
  cut,
  alpha = 0.025,
  beta = NA,
  N = NA,
  r = 1,
  maxN = 1e+06
)
}
\arguments{
\item{delta}{log(RR) between treatment and control groups.}

\item{lambda0}{Baseline hazard of control group.}

\item{t}{Average exposure time.}

\item{k}{The over-dispersion parameter for negative binomial distribution, which is 0 for poisson distribution. The default value is 0.}

\item{alpha}{One-sided type I error rate. The default value is 0.025.}

\item{beta}{Type II error rate. When \code{beta} is \code{NA} and \code{N} is not \code{NA}, the power will be returned.}

\item{N}{Sample size. When \code{N} is \code{NA} and \code{beta} is not \code{NA}, the sample size will be returned.}

\item{r}{Ratio of the sample sizes of the treatment group to the control group. The default value is 1.}

\item{cut}{A positive value for non-inferiority or equivalence margin. For example, if the non-inferiority margin for RR is 0.6, then \code{cut = -log(0.6)}. If the non-inferiority margin for RR is 1.3, then \code{cut = log(1.3)}.}

\item{direct}{\code{direct = 1} indicates that a larger RR is preferable, while \code{direct = -1} indicates that a smaller RR is preferable.}

\item{maxN}{Maximum possible sample size(\code{N}) in equivalence design. Default value is 1e6.}
}
\value{
A data frame containing input parameters and returned sample size or power.
}
\description{
Calculating sample size when given power or power when given sample size for count endpoints.
}
\details{
Taking the larger RR is preferable as an example. Sample size calculation is based on the following Z test, with the success criterion:

in superiority design:
\deqn{Z = \frac{\hat \delta}{\sqrt{\frac{\hat \sigma_0^2}{N / (r + 1)} + \frac{\hat \sigma_1^2}{Nr / (r + 1)}}}  \Phi^{-1}(1 - \alpha)}

in non-inferiority design:
\deqn{Z = \frac{\hat \delta + \Delta}{\sqrt{\frac{\hat \sigma_0^2}{N / (r + 1)} + \frac{\hat \sigma_1^2}{Nr / (r + 1)}}}  \Phi^{-1}(1 - \alpha)}

in equivalence design:
\deqn{Z_u = \frac{\hat \delta + \Delta}{\sqrt{\frac{\hat \sigma_0^2}{N / (r + 1)} + \frac{\hat \sigma_1^2}{Nr / (r + 1)}}}  \Phi^{-1}(1 - \alpha)\text{ and } Z_l = \frac{\hat \delta - \Delta}{\sqrt{\frac{\hat \sigma_0^2}{N / (r + 1)} + \frac{\hat \sigma_1^2}{Nr / (r + 1)}}} < \Phi^{-1}(\alpha)}

Where \eqn{\hat \delta = log(\hat{RR})} between treatment and control groups, \eqn{\hat \sigma_0^2 = \frac{1}{\hat \lambda_0t} + \hat k, \sigma_1^2 = \frac{1}{e^{\hat \delta}\hat \lambda_0t} + \hat k}, and \eqn{\Delta} is the non-inferiority or equivalence margin (\code{cut}).
}
\examples{
(v <- getN_Count_Super(
  delta = log(1.2),
  lambda0 = 0.5, t = 5, k = 0, alpha = 0.025, beta = 0.2, N = NA, r = 1
))
getN_Count_Super(
  delta = log(1.2),
  lambda0 = 0.5, t = 5, k = 0, alpha = 0.025, beta = NA, N = v$N, r = 1
)

(v <- getN_Count_Noninf(
  delta = log(1.1),
  lambda0 = 0.1, t = 5, k = 1, cut = log(1.4),
  alpha = 0.025, beta = 0.2, N = NA, r = 1, direct = -1
))
getN_Count_Noninf(
  delta = log(1.1),
  lambda0 = 0.1, t = 5, k = 1, cut = log(1.4),
  alpha = 0.025, beta = NA, N = v$N, r = 1, direct = -1
)
}
