% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b4_getN_Count_Super_JM1.R
\name{getN_Count_JM1}
\alias{getN_Count_JM1}
\alias{getN_Count_Super_JM1}
\alias{getN_Count_Noninf_JM1}
\alias{getN_Count_Equi_JM1}
\title{Regional sample size allocation using Japan's Method 1 for count endpoints}
\usage{
getN_Count_Super_JM1(
  delta_a,
  delta_j,
  lambda0_a,
  lambda0_j,
  t,
  k = 0,
  pi = 0.5,
  alpha = NA,
  beta = NA,
  beta1 = 0.2,
  N = NA,
  r = 1
)

getN_Count_Noninf_JM1(
  delta_a,
  delta_j,
  lambda0_a,
  lambda0_j,
  t,
  k = 0,
  pi = 0.5,
  cut,
  alpha = NA,
  beta = NA,
  beta1 = 0.2,
  N = NA,
  r = 1,
  direct = 1
)

getN_Count_Equi_JM1(
  delta_a,
  delta_j,
  lambda0_a,
  lambda0_j,
  t,
  k = 0,
  pi = 0.5,
  cut,
  alpha = NA,
  beta = NA,
  beta1 = 0.2,
  N = NA,
  r = 1,
  maxN = 1e+06
)
}
\arguments{
\item{delta_a}{log(RR) between treatment and control groups globally.}

\item{delta_j}{log(RR) between treatment and control groups for target region.}

\item{lambda0_a}{Baseline hazard of control group globally.}

\item{lambda0_j}{Baseline hazard of control group for target region.}

\item{t}{Average exposure time.}

\item{k}{The over-dispersion parameter for negative binomial distribution, which is 0 for poisson distribution. The default value is 0.}

\item{pi}{Proportion of global efficacy to retain. The default value is 0.5, which means retaining half of the efficacy.}

\item{alpha}{One-sided type I error rate for global success, which is used to calculate global sample size only when \code{N} is \code{NA}. The default value is 0.025.}

\item{beta}{Type II error rate for global success, which is used to calculate global sample size only when \code{N} is \code{NA}.}

\item{beta1}{Type II error rate for efficacy consistency between target region and globally. The default value is 0.2.}

\item{N}{Global sample size. When \code{N} is \code{NA} and \code{alpha} and \code{beta} are not \code{NAs}, \code{N} will be calculated automatically.}

\item{r}{Ratio of the sample sizes of the treatment group to the control group. The default value is 1.}

\item{cut}{A positive value for non-inferiority or equivalence margin. For example, if the non-inferiority margin for RR is 0.6, then \code{cut = -log(0.6)}. If the non-inferiority margin for RR is 1.3, then \code{cut = log(1.3)}.}

\item{direct}{\code{direct = 1} indicates that a larger RR is preferable, while \code{direct = -1} indicates that a smaller RR is preferable.}

\item{maxN}{Maximum possible sample size (\code{N}) in equivalence design. Default value is 1e6.}
}
\value{
A data frame where \code{f} is required proportion of sample size allocated to the target region, and \code{Nj} is required sample size for the target region, calculated as \code{Nj = N * f}.
}
\description{
Based on Japan's Method 1, given the global sample size and marginal probability (power) of efficacy consistency between target region and globally, calculate the required sample size allocated to the target region, in clinical trials using superiority, non-inferiority, and equivalence designs with count endpoints.
}
\details{
The global success criterion and the efficacy consistency criterion between target region and globally could be found in \code{\link{getPwr_Count_Super_JM1}}.
}
\examples{
getN_Count_Super_JM1(
  delta_a = log(1.4),
  delta_j = log(1.3),
  lambda0_a = 0.1, lambda0_j = 0.1, t = 5, k = 0, pi = 0.5, beta1 = 0.2,
  N = 300, r = 1
)

getN_Count_Noninf_JM1(
  delta_a = log(1.0),
  delta_j = log(1.1),
  lambda0_a = 0.1, lambda0_j = 0.1, t = 5, k = 0, pi = 0.5, cut = log(1.3),
  alpha = 0.025, beta = 0.2, beta1 = 0.2, N = NA, r = 1, direct = -1
)
}
\references{
1. Quan H, Li M, Chen J, et al. Assessment of Consistency of Treatment Effects in Multiregional Clinical Trials. Drug Information J. 2010;44(5):617-632. doi:10.1177/009286151004400509

2. Liao JJZ, Yu Z, Li Y. Sample size allocation in multiregional equivalence studies. Pharm Stat. 2018;17(5):570-577. doi:10.1002/pst.1871
}
\seealso{
\code{\link{getPwr_Count_Super_JM1}}, \code{\link{getN_Count_Super}}.
}
