% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3_getN_Surv_Super.R
\name{getN_Surv}
\alias{getN_Surv}
\alias{getN_Surv_Super}
\alias{getN_Surv_Noninf}
\alias{getN_Surv_Equi}
\title{Sample size and power for survival endpoints}
\usage{
getN_Surv_Super(delta, alpha = 0.025, beta = NA, N = NA, r = 1, criterion = 1)

getN_Surv_Noninf(
  delta,
  cut,
  alpha = 0.025,
  beta = NA,
  N = NA,
  r = 1,
  criterion = 1,
  direct = 1
)

getN_Surv_Equi(
  delta,
  cut,
  alpha = 0.025,
  beta = NA,
  N = NA,
  r = 1,
  criterion = 1,
  maxN = 1e+06
)
}
\arguments{
\item{delta}{log(HR) between treatment and control groups.}

\item{alpha}{One-sided type I error rate. The default value is 0.025.}

\item{beta}{Type II error rate. When \code{beta} is \code{NA} and \code{N} is not \code{NA}, the power will be returned.}

\item{N}{Sample size. When \code{N} is \code{NA} and \code{beta} is not \code{NA}, the sample size will be returned.}

\item{r}{Ratio of the sample sizes of the treatment group to the control group. The default value is 1.}

\item{criterion}{If \code{criterion = 1}, the success criterion defined on the log(HR) scale will be used. If \code{criterion = 2}, the success criterion defined on the HR scale will be used. See \code{details} for more information.}

\item{cut}{A positive value for non-inferiority or equivalence margin. For example, if the non-inferiority margin for HR is 0.6, then \code{cut = -log(0.6)}. If the non-inferiority margin for HR is 1.3, then \code{cut = log(1.3)}.}

\item{direct}{\code{direct = 1} indicates that a larger HR is preferable, while \code{direct = -1} indicates that a smaller HR is preferable.}

\item{maxN}{Maximum possible sample size(\code{N}) in equivalence design. Default value is 1e6.}
}
\value{
A data frame containing input parameters and returned sample size or power.
}
\description{
Calculating sample size when given power or power when given sample size for survival endpoints.
}
\details{
Taking the larger HR is preferable as an example. Sample size calculation is based on the following Z test, with the success criterion:

\code{criterion = 1} in superiority design:
\deqn{Z = \frac{\hat \delta}{\sqrt{\frac{1}{N / (r + 1)} + \frac{1}{Nr / (r + 1)}}} > \Phi^{-1}(1 - \alpha)}

\code{criterion = 2} in superiority design:
\deqn{Z = \frac{1 - e^{\hat \delta}}{e^{\hat \delta}\sqrt{\frac{1}{N / (r + 1)} + \frac{1}{Nr / (r + 1)}}} > \Phi^{-1}(\alpha)}

\code{criterion = 1} in non-inferiority design:
\deqn{Z = \frac{\hat \delta + \Delta}{\sqrt{\frac{1}{N / (r + 1)} + \frac{1}{Nr / (r + 1)}}} > \Phi^{-1}(1 - \alpha)}

\code{criterion = 2} in non-inferiority design:
\deqn{Z = \frac{\frac{1}{e^{\Delta}} - e^{\hat \delta}}{e^{\hat \delta}\sqrt{\frac{1}{N / (r + 1)} + \frac{1}{Nr / (r + 1)}}} > \Phi^{-1}(\alpha)}

\code{criterion = 1} in equivalence design:
\deqn{Z_u = \frac{\hat \delta + \Delta}{\sqrt{\frac{1}{N / (r + 1)} + \frac{1}{Nr / (r + 1)}}} > \Phi^{-1}(1 - \alpha) \text{ and } Z_l = \frac{\hat \delta - \Delta}{\sqrt{\frac{1}{N / (r + 1)} + \frac{1}{Nr / (r + 1)}}} > \Phi^{-1}(\alpha)}

\code{criterion = 2} in equivalence design:
\deqn{Z_l = \frac{\frac{1}{e^{\Delta}} - e^{\hat \delta}}{e^{\hat \delta}\sqrt{\frac{1}{N / (r + 1)} + \frac{1}{Nr / (r + 1)}}} > \Phi^{-1}(\alpha) \text{ and } Z_u = \frac{e^{\Delta} - e^{\hat \delta}}{e^{\hat \delta}\sqrt{\frac{1}{N / (r + 1)} + \frac{1}{Nr / (r + 1)}}} > \Phi^{-1}(1 - \alpha)}

Where \eqn{\hat \delta = log({\hat{HR}})} between treatment and control groups, and \eqn{\Delta} is the non-inferiority or equivalence margin (\code{cut}).

For \code{criterion = 2}, by delta method, \eqn{e^{\hat \delta}} approximately follows a distribution of \eqn{N(e^{\delta}, (e^{\delta}\sqrt{\frac{1}{N / (r + 1)} + \frac{1}{Nr / (r + 1)}})^2)}
}
\examples{
(v <- getN_Surv_Super(
  delta = log(1.2),
  alpha = 0.025, beta = 0.2, N = NA, r = 1, criterion = 1
))
getN_Surv_Super(
  delta = log(1.2),
  alpha = 0.025, beta = NA, N = v$N, r = 1, criterion = 1
)

(v <- getN_Surv_Noninf(
  delta = log(1.1),
  cut = log(1.3),
  alpha = 0.025, beta = 0.2, N = NA, r = 1, criterion = 2, direct = -1
))
getN_Surv_Noninf(
  delta = log(1.1),
  cut = log(1.3),
  alpha = 0.025, beta = NA, N = v$N, r = 1, criterion = 2, direct = -1
)
}
