% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/38 getN_Surv_Noninf.R
\name{getN_Surv_Noninf}
\alias{getN_Surv_Noninf}
\title{Title}
\usage{
getN_Surv_Noninf(delta, cut, alpha, beta, N, r, direct = 1)
}
\arguments{
\item{direct}{}
}
\description{
Title
}
\examples{
(v <- getN_Surv_Noninf(
  delta = log(1.1), cut = log(1.2), alpha = 0.025,
  beta = 0.2, N = NA, r = 1
))
getN_Surv_Noninf(
  delta = log(1.1), cut = log(1.2), alpha = 0.025,
  beta = NA, N = v$N, r = 1
)
}
