% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/37 getN_Surv_Super.R
\name{getN_Surv_Super}
\alias{getN_Surv_Super}
\title{Title}
\usage{
getN_Surv_Super(delta, alpha, beta, N, r)
}
\arguments{
\item{r}{}
}
\description{
Title
}
\examples{
(v <- getN_Surv_Super(
  delta = log(0.8), alpha = 0.025,
  beta = 0.2, N = NA, r = 1
))
getN_Surv_Super(
  delta = log(0.8), alpha = 0.025,
  beta = NA, N = v$N, r = 1
)
}
