% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/57 getN_Surv_Super_JM1.R
\name{getN_Surv_Super_JM1}
\alias{getN_Surv_Super_JM1}
\title{Title}
\usage{
getN_Surv_Super_JM1(delta_j, delta_nj, pi, beta1, N, r, criterion, direct = -1)
}
\arguments{
\item{direct}{}
}
\description{
Title
}
\examples{
getN_Surv_Super_JM1(
  delta_j = log(0.8), delta_nj = log(0.7),
  pi = 0.5, beta1 = 0.2, N = seq(400, 800, 200),
  criterion = c(1, 2), r = 1, direct = -1
)
}
