% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3_getNe_Surv_Super.R
\name{getNe_Surv}
\alias{getNe_Surv}
\alias{getNe_Surv_Super}
\alias{getNe_Surv_Noninf}
\alias{getNe_Surv_Equi}
\title{Number of events and power for survival endpoints}
\usage{
getNe_Surv_Super(delta, alpha = 0.025, beta = NA, Ne = NA, r = 1)

getNe_Surv_Noninf(
  delta,
  cut,
  alpha = 0.025,
  beta = NA,
  Ne = NA,
  r = 1,
  direct = 1
)

getNe_Surv_Equi(
  delta,
  cut,
  alpha = 0.025,
  beta = NA,
  Ne = NA,
  r = 1,
  maxNe = 1e+06
)
}
\arguments{
\item{delta}{A vector. log(HR) between treatment and control groups.}

\item{alpha}{A vector. One-sided type I error rate. Default value is 0.025.}

\item{beta}{A vector. Type II error rate. When \code{beta} is \code{NA} and \code{Ne} is not \code{NA}, the power will be returned.}

\item{Ne}{A vector. Number of events. When \code{Ne} is \code{NA} and \code{beta} is not \code{NA}, the number of events will be returned.}

\item{r}{A vector. Ratio of number of events of the treatment group to the control group. Default value is 1 which under H0 assumption.}

\item{cut}{A vector. Positive value for non-inferiority or equivalence margin. For example, if the non-inferiority margin for HR is 0.6, then \code{cut = -log(0.6)}. If the non-inferiority margin for HR is 1.3, then \code{cut = log(1.3)}.}

\item{direct}{\code{direct = 1} indicates that a larger HR is preferable, while \code{direct = -1} indicates that a smaller HR is preferable.}

\item{maxNe}{Maximum possible number of events (\code{Ne}) in equivalence design. Default value is 1e+06.}
}
\value{
A data frame containing input parameters and returned number of events or power.
}
\description{
Calculating required number of events when given power or power when given number of events for survival endpoints.
}
\details{
Taking the larger HR is preferable as an example. Number of events calculation is based on the following Z test, with the success criterion:

in superiority design:
\deqn{Z = \frac{\hat \delta}{\sqrt{\frac{1}{N_e / (r + 1)} + \frac{1}{N_e r / (r + 1)}}} > \Phi^{-1}(1 - \alpha)}

in non-inferiority design:
\deqn{Z = \frac{\hat \delta + \Delta}{\sqrt{\frac{1}{N_e / (r + 1)} + \frac{1}{N_e r / (r + 1)}}} > \Phi^{-1}(1 - \alpha)}

in equivalence design:
\deqn{Z_u = \frac{\hat \delta + \Delta}{\sqrt{\frac{1}{N_e / (r + 1)} + \frac{1}{N_e r / (r + 1)}}} > \Phi^{-1}(1 - \alpha) \text{ and } Z_l = \frac{\hat \delta - \Delta}{\sqrt{\frac{1}{N_e / (r + 1)} + \frac{1}{N_e r / (r + 1)}}} < \Phi^{-1}(\alpha)}

Where \eqn{\hat \delta = log({\hat{HR}})} between treatment and control groups, and \eqn{\Delta} is the non-inferiority or equivalence margin (\code{cut}).
}
\examples{
(v <- getNe_Surv_Super(
  delta = log(1.2),
  alpha = 0.025, beta = 0.2, Ne = NA, r = 1
))
getNe_Surv_Super(
  delta = log(1.2),
  alpha = 0.025, beta = NA, Ne = v$Ne, r = 1
)

(v <- getNe_Surv_Noninf(
  delta = log(1.1),
  cut = log(1.3),
  alpha = 0.025, beta = 0.2, Ne = NA, r = 1, direct = -1
))
getNe_Surv_Noninf(
  delta = log(1.1),
  cut = log(1.3),
  alpha = 0.025, beta = NA, Ne = v$Ne, r = 1, direct = -1
)
}
