% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3_getPwr_Bin_Super_JM1.R
\name{getPwr_Bin_JM1}
\alias{getPwr_Bin_JM1}
\alias{getPwr_Bin_Super_JM1}
\alias{getPwr_Bin_Noninf_JM1}
\alias{getPwr_Bin_Equi_JM1}
\title{Power of mRCT using Japan's Method 1 for binary endpoints}
\usage{
getPwr_Bin_Super_JM1(
  p1_j,
  p0_j,
  p1_nj = NA,
  p0_nj = NA,
  p1_a = NA,
  p0_a = NA,
  f,
  pi = 0.5,
  alpha = 0.025,
  beta = NA,
  N = NA,
  r = 1,
  scale = "RD",
  sim = FALSE,
  nsim = 1000,
  seed = 0,
  numcore = 4
)

getPwr_Bin_Noninf_JM1(
  p1_j,
  p0_j,
  p1_nj = NA,
  p0_nj = NA,
  p1_a = NA,
  p0_a = NA,
  f,
  pi = 0.5,
  cut,
  alpha = 0.025,
  beta = NA,
  N = NA,
  r = 1,
  scale = "RD",
  direct = 1,
  sim = FALSE,
  nsim = 1000,
  seed = 0,
  numcore = 4
)

getPwr_Bin_Equi_JM1(
  p1_j,
  p0_j,
  p1_nj = NA,
  p0_nj = NA,
  p1_a = NA,
  p0_a = NA,
  f,
  pi = 0.5,
  cut,
  alpha,
  beta = NA,
  N = NA,
  r = 1,
  scale = "RD",
  sim = FALSE,
  nsim = 1000,
  seed = 0,
  numcore = 4
)
}
\arguments{
\item{p1_j}{Rate of treatment group in target region.}

\item{p0_j}{Rate of control group in target region.}

\item{p1_nj}{Rate of treatment group in other regions. When \code{p1_nj} is not \code{NA}, \code{p1_a} will be calculated automatically.}

\item{p0_nj}{Rate of control group in other regions. When \code{p0_nj} is not \code{NA}, \code{p0_a} will be calculated automatically.}

\item{p1_a}{Rate of treatment group globally.}

\item{p0_a}{Rate of the control group globally.}

\item{f}{Proportion of sample size allocated to target region.}

\item{pi}{Proportion of global efficacy to retain. The default value is 0.5, which means retaining half of the efficacy.}

\item{alpha}{One-sided type I error rate for global success. The default value is 0.025.}

\item{beta}{Type II error rate for global success, which is used to calculate global sample size only when \code{N} is \code{NA}.}

\item{N}{Global sample size. When N is \code{NA} and \code{alpha} and \code{beta} are not \code{NA}, \code{N} will be calculated automatically.}

\item{r}{Ratio of the sample sizes of the treatment group to the control group. The default value is 1.}

\item{scale}{Optional values are "RD" for rate difference, "RR" for relative risk, and "OR" for odds ratio. Default value is "RD".}

\item{sim}{Logical value. When set to \code{FALSE}, theoretical calculation is performed. When set to \code{TRUE}, simulation is used, which is more time-consuming.}

\item{nsim}{Number of simulations.}

\item{seed}{Random seed for simulation.}

\item{numcore}{Number of CPU cores to use during simulation.}

\item{cut}{A positive value for non-inferiority or equivalence margin. For RD, the margin is on the original scale. For RR and OR, the margin is on the log scale. For example, if the non-inferiority margins for RD, RR, OR are 0.2, 0.6, and 1.3, then the \code{cut = 0.2}, \code{cut = -log(0.6)}, and \code{cut = log(1.3)}, respectively.}

\item{direct}{If \code{direct = 1}, larger values of RD/RR/OR are preferable. If \code{direct = -1}, smaller values of RD/RR/OR are preferable.}
}
\value{
A data frame containing input parameters and returned power.
\describe{
  \item{\code{pwr1}}{The marginal probability of global success.}
  \item{\code{pwr2}}{The marginal probability that the target region efficacy is consistent with the global efficacy.}
  \item{\code{pwr3}}{The conditional probability that the target region efficacy is consistent with the global efficacy given global success.}
  \item{\code{pwr4}}{The joint probability of global success and the target region efficacy being consistent with the global efficacy.}
}
}
\description{
Based on Japan's Method 1, given the global and target region sample sizes, calculate and simulate the marginal probabilities, conditional probabilities, and joint probabilities of global success and efficacy consistency between target region and globally, in clinical trials using superiority, non-inferiority, and equivalence designs with binary endpoints.
}
\details{
Taking the larger RD/RR/OR is preferable as an example. The global success criterion and the efficacy consistency criterion between target region and globally

in superiority design:
\deqn{Z_a = \frac{\hat \delta_a}{\sqrt{Var(\hat \delta_a)}}> \Phi^{-1}(1 - \alpha)}
\deqn{\hat \delta_j - \pi\hat \delta_a > 0}

in non-inferiority design:
\deqn{Z_a = \frac{\hat \delta_a + \Delta}{\sqrt{Var(\hat \delta_a)}} > \Phi^{-1}(1 - \alpha)}
\deqn{\hat \delta_j - \hat \delta_a + \pi\Delta > 0}

in equivalence design:
\deqn{Z_{a_u} = \frac{\hat \delta_a + \Delta}{\sqrt{Var(\hat \delta_a)}} > \Phi^{-1}(1 - \alpha)\text{ and }Z_{a_l} = \frac{\hat \delta_a - \Delta}{\sqrt{Var(\hat \delta_a)}} < \Phi^{-1}(\alpha)}
\deqn{\hat \delta_j - \hat \delta_a + \pi\Delta > 0\text{ and }\hat \delta_j - \hat \delta_a - \pi\Delta < 0}

Where \eqn{\hat \delta = \hat p_1 - \hat p_0} for RD, \eqn{\hat \delta = log(\frac{\hat p_1}{\hat p_0})} for RR, and  \eqn{\hat \delta = log(\frac{\hat p_1 / (1 - \hat p_1)}{\hat p_0 / (1 - \hat p_0)})}  for OR.  \eqn{\Delta} is the non-inferiority or equivalence margin (\code{cut}).
}
\examples{
getPwr_Bin_Super_JM1(
  p1_j = 0.7, p0_j = 0.5, p1_a = 0.75, p0_a = 0.5, f = seq(0.1, 0.9, 0.1),
  pi = 0.5, alpha = 0.025, beta = NA, N = 200, r = 1, scale = "RD", sim = FALSE
)

# p1_a and p0_a will be calculated based on p1_j and p1_nj, p0_j and p0_nj, respectively.
# Global sample size will be calculated based on beta.
getPwr_Bin_Noninf_JM1(
  p1_j = 0.6, p0_j = 0.5, p1_nj = 0.5, p0_nj = 0.5, f = seq(0.1, 0.9, 0.1),
  pi = 0.5, cut = log(1.4),
  alpha = 0.025, beta = 0.2, N = NA, r = 1, scale = "RR", direct = -1,
  sim = FALSE
)
}
\references{
1. Quan H, Li M, Chen J, et al. Assessment of Consistency of Treatment Effects in Multiregional Clinical Trials. Drug Information J. 2010;44(5):617-632. doi:10.1177/009286151004400509

2. Liao JJZ, Yu Z, Li Y. Sample size allocation in multiregional equivalence studies. Pharm Stat. 2018;17(5):570-577. doi:10.1002/pst.1871
}
