% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a4_getPwr_Bin_Super_JM2.R
\name{getPwr_Bin_JM2}
\alias{getPwr_Bin_JM2}
\alias{getPwr_Bin_Super_JM2}
\alias{getPwr_Bin_Noninf_JM2}
\alias{getPwr_Bin_Equi_JM2}
\title{Power of mRCT using Japan's Method 2 for binary endpoints}
\usage{
getPwr_Bin_Super_JM2(
  p1_i,
  p0_i,
  fi,
  alpha = 0.025,
  beta = NA,
  N = NA,
  r = 1,
  scale = "RD",
  sim = FALSE,
  nsim = 1000,
  seed = 0
)

getPwr_Bin_Noninf_JM2(
  p1_i,
  p0_i,
  fi,
  cut,
  alpha = 0.025,
  beta = NA,
  N = NA,
  r = 1,
  scale = "RD",
  direct = 1,
  sim = FALSE,
  nsim = 1000,
  seed = 0
)

getPwr_Bin_Equi_JM2(
  p1_i,
  p0_i,
  fi,
  cut,
  alpha = 0.025,
  beta = NA,
  N = NA,
  r = 1,
  scale = "RD",
  sim = FALSE,
  nsim = 1000,
  seed = 0
)
}
\arguments{
\item{p1_i}{A vector. Rate of treatment group in each region.}

\item{p0_i}{A vector. Rate of control group in each region.}

\item{fi}{A vector. Proportion of sample size allocated to each region.}

\item{alpha}{One-sided type I error rate for global success. The default value is 0.025.}

\item{beta}{Type II error rate for global success, which is used to calculate global sample size only when \code{N} is \code{NA}.}

\item{N}{Global sample size. Global sample size. When \code{N} is \code{NA} and \code{beta} is not \code{NA}, \code{N} will be calculated automatically.}

\item{r}{Ratio of the sample sizes of the treatment group to the control group. The default value is 1.}

\item{scale}{Optional values are "RD" for rate difference, "RR" for relative risk, and "OR" for odds ratio. Default value is "RD".}

\item{sim}{Logical value. When set to \code{FALSE}, theoretical calculation is performed. When set to \code{TRUE}, simulation is used, which is more time-consuming.}

\item{nsim}{Number of simulations.}

\item{seed}{Random seed for simulation.}

\item{cut}{A positive value for non-inferiority or equivalence margin. For RD, the margin is on the original scale. For RR and OR, the margin is on the log scale. For example, if the non-inferiority margins for RD, RR, and OR are 0.2, 0.6, and 1.3, then the \code{cut = 0.2}, \code{cut = -log(0.6)}, and \code{cut = log(1.3)}, respectively.}

\item{direct}{If \code{direct = 1}, larger values of RD/RR/OR are preferable. If \code{direct = -1}, smaller values of RD/RR/OR are preferable.}
}
\value{
A list where:
\itemize{
  \item{overall}{
    \itemize{
      \item{\code{pwr1 }}{The marginal probability of global success.}
      \item{\code{pwr2 }}{The marginal probability that all region's efficacy is consistent with the global efficacy.}
      \item{\code{pwr3 }}{The conditional probability that all region's efficacy is consistent with the global efficacy given global success.}
      \item{\code{pwr4 }}{The joint probability of global success and all region's efficacy being consistent with the global efficacy.}
    }
  }
  \item{\code{pwr_margin }}{The marginal probability that the ith region efficacy is consistent with the global efficacy.}
  \item{\code{pwr_condition }}{The conditional probability that the ith region efficacy is consistent with the global efficacy given global success.}
  \item{\code{pwr_joint }}{The joint probability of global success and the ith region efficacy being consistent with the global efficacy.}
}
}
\description{
Based on Japan's Method 2, given the global and target region sample sizes, calculate and simulate the marginal probabilities, conditional probabilities, and joint probabilities of global success and efficacy consistency between target region and globally, in clinical trials using superiority, non-inferiority, and equivalence designs with binary endpoints.
}
\details{
Taking the larger RD/RR/OR is preferable as an example. The global success criterion and the efficacy consistency criterion between target region and globally

in superiority design:
\deqn{Z_a = \frac{\hat \delta_a}{\sqrt{Var(\hat \delta_a)}} > \Phi^{-1}(1 - \alpha)}
\deqn{\hat \delta_i > 0 \text{ for i = 1, 2, .., m}}

in non-inferiority design:
\deqn{Z_a = \frac{\hat \delta_a + \Delta}{\sqrt{Var(\hat \delta_a)}} > \Phi^{-1}(1 - \alpha)}
\deqn{\hat \delta_i+ \Delta > 0 \text{ for i = 1, 2, .., m}}

in equivalence design:
\deqn{Z_{a_u} = \frac{\hat \delta_a + \Delta}{\sqrt{Var(\hat \delta_a)}} > \Phi^{-1}(1 - \alpha)\text{ and }Z_{a_l} = \frac{\hat \delta_a - \Delta}{\sqrt{Var(\hat \delta_a)}} < \Phi^{-1}(\alpha)}
\deqn{\hat \delta_i + \Delta> 0 \text{ and } \hat \delta_i - \Delta < 0 \text{ for i = 1, 2, .., m}}

Where \eqn{\hat \delta = \hat p_1 - \hat p_0} for RD, \eqn{\hat \delta = log(\frac{\hat p_1}{\hat p_0})} for RR, and  \eqn{\hat \delta = log(\frac{\hat p_1 / (1 - \hat p_1)}{\hat p_0 / (1 - \hat p_0)})}  for OR.  \eqn{\Delta} is the non-inferiority or equivalence margin (\code{cut}).
}
\examples{
f_set <- seq(0.1, 0.9, 0.1)
map_dfr(
  .x = 1:length(f_set),
  .f = function(i) {
    f <- f_set[i]
    res <- getPwr_Bin_Super_JM2(
      p1_i = c(0.7, 0.75),
      p0_i = c(0.5, 0.5),
      fi = c(f, 1 - f),
      alpha = 0.025, beta = NA, N = 100, r = 1, scale = "RD", sim = FALSE
    )$overall
    res$f <- f
    res
  }
)

# Global rates of treatment and control groups will be calculated based on p1_i, p0_i, and fi.
# Global sample size will be calculated based on beta.
f_set <- seq(0.1, 0.9, 0.1)
map_dfr(
  .x = 1:length(f_set),
  .f = function(i) {
    f <- f_set[i]
    res <- getPwr_Bin_Noninf_JM2(
      p1_i = c(0.6, 0.5),
      p0_i = c(0.5, 0.5),
      fi = c(f, 1 - f),
      cut = log(1.4),
      alpha = 0.025, beta = 0.2, N = NA, r = 1, scale = "RR", direct = -1,
      sim = FALSE
    )$overall
    res$f <- f
    res
  }
)
}
\references{
1. Quan H, Li M, Chen J, et al. Assessment of Consistency of Treatment Effects in Multiregional Clinical Trials. Drug Information J. 2010;44(5):617-632. doi:10.1177/009286151004400509

2. Liao JJZ, Yu Z, Li Y. Sample size allocation in multiregional equivalence studies. Pharm Stat. 2018;17(5):570-577. doi:10.1002/pst.1871
}
