% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11 getPwr_Bin_Noninf_JM2.R
\name{getPwr_Bin_Noninf_JM2}
\alias{getPwr_Bin_Noninf_JM2}
\title{Title}
\usage{
getPwr_Bin_Noninf_JM2(
  pt_i,
  pc_i,
  fi,
  cut,
  alpha,
  N,
  r,
  direct = 1,
  sim = FALSE,
  nsim = 1000,
  seed = 0
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
\examples{
f_set <- seq(0.1, 0.9, 0.1)
map_dfr(.x = 1:length(f_set), .f = function(i) {
  f <- f_set[i]
  res <- getPwr_Bin_Noninf_JM2(
    pt_i = c(0.5, 0.6),
    pc_i = c(0.6, 0.6),
    fi = c(f, 1 - f), cut = 0.3,
    alpha = 0.025, N = 100, r = 1, direct = 1, sim = FALSE
  )$overall
  res$M <- "calc"
  res$f <- f
  res
})
}
