% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3 getPwr_Con_Equi_JM1.R
\name{getPwr_Con_Equi_JM1}
\alias{getPwr_Con_Equi_JM1}
\title{Title}
\usage{
getPwr_Con_Equi_JM1(
  delta_j,
  delta_nj,
  sigma,
  f,
  pi,
  cut,
  alpha,
  N,
  r,
  sim = F,
  nsim = 1000,
  seed = 0
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
\examples{
library(ggplot2)
dat1 <- getPwr_Con_Equi_JM1(
  delta_j = -0.2, delta_nj = -0.1, sigma = 1,
  f = seq(0.1, 0.9, 0.1), pi = 0.5, cut = 0.4, alpha = 0.025,
  N = 400, r = 1, sim = F
)
dat1$M <- "calc"
dat2 <- getPwr_Con_Equi_JM1(
  delta_j = -0.2, delta_nj = -0.1, sigma = 1,
  f = seq(0.1, 0.9, 0.1), pi = 0.5, cut = 0.4, alpha = 0.025,
  N = 400, r = 1, sim = T
)
dat2$M <- "sim"

dat <- bind_rows(dat1, dat2)
dat <- gather(
  data = dat, key = "pwr class", value = "pwr",
  p1, p2, p3, p4
)

ggplot(data = dat, aes(x = f, y = pwr, linetype = M)) +
  geom_point() +
  geom_line() +
  facet_wrap(vars(`pwr class`), nrow = 2) +
  scale_x_continuous(breaks = seq(0.1, 0.9, 0.2)) +
  scale_y_continuous(limits = c(0, 1), breaks = seq(0, 1.0, 0.2)) +
  labs(x = "allocation ratio", color = "", linetype = "")
}
