% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6 getPwr_Con_Equi_JM2.R
\name{getPwr_Con_Equi_JM2}
\alias{getPwr_Con_Equi_JM2}
\title{Title}
\usage{
getPwr_Con_Equi_JM2(
  delta_i,
  sigma,
  fi,
  cut,
  alpha,
  N,
  r,
  sim = FALSE,
  nsim = 1000,
  seed = 0
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
\examples{
f_set <- seq(0.1, 0.9, 0.1)
map_dfr(.x = 1:length(f_set), .f = function(i) {
  f <- f_set[i]
  res <- getPwr_Con_Equi_JM2(
    delta_i = c(-0.5, 0), sigma = 4,
    fi = c(f, 1 - f), cut = 2,
    alpha = 0.025, N = 200, r = 1, sim = FALSE
  )$overall
  res$M <- "calc"
  res$f <- f
  res
})
}
