% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a1_getPwr_Con_Super_JM1.R
\name{getPwr_Con_JM1}
\alias{getPwr_Con_JM1}
\alias{getPwr_Con_Super_JM1}
\alias{getPwr_Con_Noninf_JM1}
\alias{getPwr_Con_Equi_JM1}
\title{Power of mRCT using Japan's Method 1 for continuous endpoints}
\usage{
getPwr_Con_Super_JM1(
  delta_j,
  delta_nj = NA,
  delta_a = NA,
  sigma,
  f,
  pi = 0.5,
  alpha = 0.025,
  beta = NA,
  N = NA,
  r = 1,
  sim = FALSE,
  nsim = 1000,
  seed = 0,
  numcore = 4
)

getPwr_Con_Noninf_JM1(
  delta_j,
  delta_nj = NA,
  delta_a = NA,
  sigma,
  f,
  pi = 0.5,
  cut,
  alpha = 0.025,
  beta = NA,
  N = NA,
  r = 1,
  direct = 1,
  sim = FALSE,
  nsim = 1000,
  seed = 0,
  numcore = 4
)

getPwr_Con_Equi_JM1(
  delta_j,
  delta_nj = NA,
  delta_a = NA,
  sigma,
  f,
  pi = 0.5,
  cut,
  alpha = 0.025,
  beta = NA,
  N = NA,
  r = 1,
  sim = FALSE,
  nsim = 1000,
  seed = 0,
  numcore = 4
)
}
\arguments{
\item{delta_j}{Mean difference between treatment and control groups in target region.}

\item{delta_nj}{Mean difference between treatment and control groups in other regions. When \code{delta_nj} is not \code{NA}, \code{delta_a} will be calculated automatically.}

\item{delta_a}{Mean difference between treatment and control groups globally.}

\item{sigma}{Common standard deviation.}

\item{f}{Proportion of sample size allocated to target region.}

\item{pi}{Proportion of global efficacy to retain. The default value is 0.5, which means retaining half of the efficacy.}

\item{alpha}{One-sided type I error rate for global success. The default value is 0.025.}

\item{beta}{Type II error rate for global success, which is used to calculate global sample size only when \code{N} is \code{NA}.}

\item{N}{Global sample size. When \code{N} is \code{NA} and \code{beta} is not \code{NA}, \code{N} will be calculated automatically.}

\item{r}{Ratio of the sample sizes of the treatment group to the control group. The default value is 1.}

\item{sim}{Logical value. When set to \code{FALSE}, theoretical calculation is performed. When set to \code{TRUE}, simulation is used, which is more time-consuming.}

\item{nsim}{Number of simulations.}

\item{seed}{Random seed for simulation.}

\item{numcore}{Number of CPU cores to use during simulation.}

\item{cut}{A positive value for non-inferiority or equivalence margin.}

\item{direct}{\code{direct = 1} indicates that a larger mean difference is preferable, while \code{direct = -1} indicates that a smaller mean difference is preferable.}
}
\value{
A data frame containing input parameters and returned power.
\describe{
  \item{\code{pwr1 }}{The marginal probability of global success.}
  \item{\code{pwr2 }}{The marginal probability that the target region efficacy is consistent with the global efficacy.}
  \item{\code{pwr3 }}{The conditional probability that the target region efficacy is consistent with the global efficacy given global success.}
  \item{\code{pwr4 }}{The joint probability of global success and the target region efficacy being consistent with the global efficacy.}
}
}
\description{
Based on Japan's Method 1, given the global and target region sample sizes, calculate and simulate the marginal probabilities, conditional probabilities, and joint probabilities of global success and efficacy consistency between target region and globally, in clinical trials using superiority, non-inferiority, and  equivalence designs with continuous endpoints.
}
\details{
Taking the larger mean difference is preferable as an example. The global success criterion and the efficacy consistency criterion between target region and globally

in superiority design:
\deqn{Z_a = \frac{\hat \delta_a}{\sqrt{Var(\hat \delta_a)}} > \Phi^{-1}(1 - \alpha)}
\deqn{\hat \delta_j - \pi\hat \delta_a > 0}

in non-inferiority design:
\deqn{Z_a = \frac{\hat \delta_a + \Delta}{\sqrt{Var(\hat \delta_a)}} > \Phi^{-1}(1 - \alpha)}
\deqn{\hat \delta_j - \hat \delta_a + \pi\Delta > 0}

in equivalence design:
\deqn{Z_{a_u} = \frac{\hat \delta_a + \Delta}{\sqrt{Var(\hat \delta_a)}} > \Phi^{-1}(1 - \alpha)\text{ and }Z_{a_l} = \frac{\hat \delta_a - \Delta}{\sqrt{Var(\hat \delta_a)}} < \Phi^{-1}(\alpha)}
\deqn{\hat \delta_j - \hat \delta_a + \pi\Delta > 0\text{ and }\hat \delta_j - \hat \delta_a - \pi\Delta < 0}

Where \eqn{\hat \delta} is the mean difference between treatment and control groups, and \eqn{\Delta} is the non-inferiority or equivalence margin.
}
\examples{
getPwr_Con_Super_JM1(
  delta_j = 0.5, delta_a = 0.7, sigma = 1, f = seq(0.1, 0.9, 0.1),
  pi = 0.5, alpha = 0.025, beta = NA, N = 100, r = 1, sim = FALSE
)

# delta_a will be calculated based on delta_j and delta_nj.
# Global sample size will be calculated based on beta.
getPwr_Con_Noninf_JM1(
  delta_j = 0.2, delta_nj = 0.1, sigma = 1, f = seq(0.1, 0.9, 0.1),
  pi = 0.5, cut = 0.4, alpha = 0.025, beta = 0.2, N = NA, r = 1, direct = -1,
  sim = FALSE
)
}
\references{
1. Quan H, Li M, Chen J, et al. Assessment of Consistency of Treatment Effects in Multiregional Clinical Trials. Drug Information J. 2010;44(5):617-632. doi:10.1177/009286151004400509

2. Liao JJZ, Yu Z, Li Y. Sample size allocation in multiregional equivalence studies. Pharm Stat. 2018;17(5):570-577. doi:10.1002/pst.1871
}
