% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1 getPwr_Con_Super_JM1.R
\name{getPwr_Con_Super_JM1}
\alias{getPwr_Con_Super_JM1}
\title{Title}
\usage{
getPwr_Con_Super_JM1(
  delta_j,
  delta_nj,
  sigma,
  f,
  pi,
  alpha,
  N,
  r,
  direct = 1,
  sim = FALSE,
  nsim = 1000,
  seed = 0
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
\examples{
getPwr_Con_Super_JM1(
  delta_j = 0.5, delta_nj = 0.7, sigma = 1,
  f = seq(0.1, 0.9, 0.1),
  pi = 0.5, alpha = 0.025, N = 100, r = 1, direct = 1, sim = FALSE
)
}
