% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a8_getPwr_Count_Super_JM2.R
\name{getPwr_Count_JM2}
\alias{getPwr_Count_JM2}
\alias{getPwr_Count_Super_JM2}
\alias{getPwr_Count_Noninf_JM2}
\alias{getPwr_Count_Equi_JM2}
\title{Power of mRCT using Japan's Method 2 for count endpoints}
\usage{
getPwr_Count_Super_JM2(
  delta_i,
  lambda0_i,
  t_i,
  k = 0,
  f_i,
  alpha = 0.025,
  beta = NA,
  N = NA,
  r = 1,
  sim = FALSE,
  nsim = 1000,
  seed = 0
)

getPwr_Count_Noninf_JM2(
  delta_i,
  lambda0_i,
  t_i,
  k = 0,
  f_i,
  cut,
  cut_i = NA,
  alpha = 0.025,
  beta = NA,
  N = NA,
  r = 1,
  direct = 1,
  sim = FALSE,
  nsim = 1000,
  seed = 0
)

getPwr_Count_Equi_JM2(
  delta_i,
  lambda0_i,
  t_i,
  k = 0,
  f_i,
  cut,
  cut_i = NA,
  alpha = 0.025,
  beta = NA,
  N = NA,
  r = 1,
  sim = FALSE,
  nsim = 1000,
  seed = 0,
  maxN = 1e+06
)
}
\arguments{
\item{delta_i}{A vector with length equal to number of regions. log(RR) between treatment and control groups for each region.}

\item{lambda0_i}{A vector with length equal to number of regions. Baseline hazard of control group for each region.}

\item{t_i}{Average exposure time for each region.}

\item{k}{The over-dispersion parameter (k > 0) for negative binomial distribution, which is 0 for poisson distribution.}

\item{f_i}{A vector with length equal to number of regions. Proportion of sample size allocated to each region.}

\item{alpha}{One-sided type I error rate for global success. Default value is 0.025.}

\item{beta}{Type II error rate for global success, which is used to calculate global sample size only when \code{N} is \code{NA}.}

\item{N}{Global sample size. Global sample size. When \code{N} is \code{NA} and \code{beta} is not \code{NA}, \code{N} will be calculated automatically.}

\item{r}{Ratio of sample sizes of treatment group to control group. Default value is 1.}

\item{sim}{Logical value. When set to \code{FALSE}, theoretical calculation is performed. When set to \code{TRUE}, simulation is used, which is more time-consuming.}

\item{nsim}{Number of simulations.}

\item{seed}{Random seed for simulation.}

\item{cut}{A positive value for non-inferiority or equivalence margin in global trial, and this margin is also used in the ith region when the ith element of cut_i is \code{NA}. For example, if the non-inferiority margin for RR is 0.6, then \code{cut = -log(0.6)}. If the non-inferiority margin for RR is 1.3, then \code{cut = log(1.3)}.}

\item{cut_i}{A vector with length equal to number of regions. Positive value for non-inferiority or equivalence margin in each region. When \code{cut_i = NA} (default), globally margin will be used for each region.}

\item{direct}{\code{direct = 1} indicates that a larger RR is preferable, while \code{direct = -1} indicates that a smaller RR is preferable.}

\item{maxN}{Maximum possible global sample size (\code{N}) in equivalence design. Default value is 1e+06.}
}
\value{
A list where:
\itemize{
  \item{overall}{
    \itemize{
      \item{\code{pwr1 }}{The marginal probability of global success.}
      \item{\code{pwr2 }}{The marginal probability that all region's efficacy is consistent with the global efficacy.}
      \item{\code{pwr3 }}{The joint probability of global success and all region's efficacy being consistent with the global efficacy.}
      \item{\code{pwr4 }}{The conditional probability that all region's efficacy is consistent with the global efficacy given global success.}
    }
  }
  \item{\code{cut_i }}{The non-inferiority or equivalence margin in each region (\code{cut_i}).}
  \item{\code{pwr_margin }}{The marginal probability that the ith region efficacy is consistent with the global efficacy.}
  \item{\code{pwr_joint }}{The joint probability of global success and the ith region efficacy being consistent with the global efficacy.}
  \item{\code{pwr_condition }}{The conditional probability that the ith region efficacy is consistent with the global efficacy given global success.}
}
}
\description{
Based on Japan's Method 2, given the global and target region sample sizes, calculate and simulate the marginal probabilities, conditional probabilities, and joint probabilities of global success and efficacy consistency between target region and globally, in clinical trials using superiority, non-inferiority, and equivalence designs with count endpoints.
}
\details{
Taking the larger RR is preferable as an example. The global success criterion and the efficacy consistency criterion between target region and globally

in superiority design:
\deqn{Z_a = \frac{\hat \delta_a}{\sqrt{Var(\hat \delta_a)}} > \Phi^{-1}(1 - \alpha)}
\deqn{\hat \delta_i > 0 \text{ for i = 1, 2, .., m}}

in non-inferiority design:
\deqn{Z_a = \frac{\hat \delta_a + \Delta}{\sqrt{Var(\hat \delta_a)}} > \Phi^{-1}(1 - \alpha)}
\deqn{\hat \delta_i + \Delta_i > 0 \text{ for i = 1, 2, .., m}}

in equivalence design:
\deqn{Z_{a_u} = \frac{\hat \delta_a + \Delta}{\sqrt{Var(\hat \delta_a)}} > \Phi^{-1}(1 - \alpha)\text{ and }Z_{a_l} = \frac{\hat \delta_a - \Delta}{\sqrt{Var(\hat \delta_a)}} < \Phi^{-1}(\alpha)}
\deqn{\hat \delta_i + \Delta_i > 0 \text{ and } \hat \delta_i - \Delta_i < 0 \text{ for i = 1, 2, .., m}}

Where \eqn{\hat \delta = log(\hat {RR})} between treatment and control groups, and \eqn{\Delta} and \eqn{\Delta_i} are the non-inferiority or equivalence margins in global trial (\code{cut}) and each region (\code{cut_i}), respectively.
}
\examples{
getPwr_Count_Super_JM2(
  delta_i = c(log(1.2), log(1.4)),
  lambda0_i = c(0.1, 0.1),
  t_i = c(5, 5), k = 0, f_i = c(0.5, 0.5),
  alpha = 0.025, beta = NA, N = 300, r = 1, sim = FALSE
)

# Global log(RR) will be calculated based on delta_i and f_i,
# and global lambda0 will be calculated based on lambda0_i and f_i
# Non-inferiority margin in global trial and each region is log(1.3).
# Global sample size will be calculated based on beta.
getPwr_Count_Noninf_JM2(
  delta_i = c(log(1.1), log(1.0)),
  lambda0_i = c(0.1, 0.1),
  t_i = c(5, 5), k = 0, f_i = c(0.5, 1 - 0.5),
  cut = log(1.3),
  alpha = 0.025, beta = 0.2, N = NA, r = 1, direct = -1, sim = FALSE
)
}
\references{
1. Quan H, Li M, Chen J, et al. Assessment of Consistency of Treatment Effects in Multiregional Clinical Trials. Drug Information J. 2010;44(5):617-632. doi:10.1177/009286151004400509

2. Liao JJZ, Yu Z, Li Y. Sample size allocation in multiregional equivalence studies. Pharm Stat. 2018;17(5):570-577. doi:10.1002/pst.1871
}
