% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15 getPwr_Surv_Equi_JM1.R
\name{getPwr_Surv_Equi_JM1}
\alias{getPwr_Surv_Equi_JM1}
\title{Title}
\usage{
getPwr_Surv_Equi_JM1(
  delta_j,
  delta_nj,
  f,
  pi,
  cut,
  alpha,
  N,
  r,
  lambda0_j = 1,
  lambda0_nj = 1,
  sim = FALSE,
  nsim = 1000,
  seed = 0
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
\examples{
getPwr_Surv_Equi_JM1(
  delta_j = log(1.1), delta_nj = log(1.0),
  f = seq(0.1, 0.9, 0.1), cut = log(1.3),
  pi = 0.5, alpha = 0.025, N = 400, r = 1, sim = FALSE
)
}
