% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17 getPwr_Surv_Noninf_JM2.R
\name{getPwr_Surv_Noninf_JM2}
\alias{getPwr_Surv_Noninf_JM2}
\title{Title}
\usage{
getPwr_Surv_Noninf_JM2(
  delta_i,
  fi,
  cut,
  alpha,
  N,
  r,
  direct = -1,
  lambda0_i = 1,
  sim = FALSE,
  nsim = 1000,
  seed = 0
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
\examples{
f_set <- seq(0.1, 0.9, 0.1)
map_dfr(.x = 1:length(f_set), .f = function(i) {
  f <- f_set[i]
  res <- getPwr_Surv_Noninf_JM2(
    delta_i = c(log(1.1), log(1.0)),
    fi = c(f, 1 - f), cut = log(1.3),
    alpha = 0.025, N = 300, r = 1, direct = -1, sim = FALSE
  )$overall
  res$M <- "calc"
  res$f <- f
  res
})
}
