% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13 getPwr_Surv_Super_JM1.R
\name{getPwr_Surv_Super_JM1}
\alias{getPwr_Surv_Super_JM1}
\title{Title}
\usage{
getPwr_Surv_Super_JM1(
  delta_j,
  delta_nj,
  f,
  pi,
  alpha,
  N,
  r,
  criterion,
  direct = -1,
  lambda0_j = 1,
  lambda0_nj = 1,
  sim = F,
  nsim = 1000,
  seed = 0
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
\examples{
library(ggplot2)
dat1 <- getPwr_Surv_Super_JM1(
  delta_j = log(0.8), delta_nj = log(0.6),
  f = seq(0.1, 0.9, 0.1),
  pi = 0.5, alpha = 0.025, N = 200, r = 1,
  criterion = 1, direct = -1, sim = F
)
dat1$M <- "calc"
dat2 <- getPwr_Surv_Super_JM1(
  delta_j = log(0.8), delta_nj = log(0.6),
  f = seq(0.1, 0.9, 0.1),
  pi = 0.5, alpha = 0.025, N = 200, r = 1,
  criterion = 1, direct = -1, sim = T
)
dat2$M <- "sim"
dat3 <- getPwr_Surv_Super_JM1(
  delta_j = log(0.8), delta_nj = log(0.6),
  f = seq(0.1, 0.9, 0.1),
  pi = 0.5, alpha = 0.025, N = 200, r = 1,
  criterion = 2, direct = -1, sim = F
)
dat3$M <- "calc"
dat4 <- getPwr_Surv_Super_JM1(
  delta_j = log(0.8), delta_nj = log(0.6),
  f = seq(0.1, 0.9, 0.1),
  pi = 0.5, alpha = 0.025, N = 200, r = 1,
  criterion = 2, direct = -1, sim = T
)
dat4$M <- "sim"

dat <- bind_rows(dat1, dat2, dat3, dat4)
dat <- gather(data = dat, key = "pwr class", value = "pwr", p1, p2, p3, p4)

ggplot(data = dat, aes(x = f, y = pwr, color = factor(criterion))) +
  geom_point() +
  geom_line(aes(linetype = M)) +
  facet_wrap(vars(`pwr class`), nrow = 2) +
  scale_x_continuous(breaks = seq(0.1, 0.9, 0.2)) +
  scale_y_continuous(limits = c(0, 1), breaks = seq(0, 1.0, 0.2)) +
  labs(x = "allocation ratio", color = "", linetype = "")
}
