% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{Ne_to_N}
\alias{Ne_to_N}
\title{Transform number of events to sample size}
\usage{
Ne_to_N(
  Ne = NA,
  r = 1,
  lambda0,
  lambda1,
  dropoutRate,
  dropoutTime = 1,
  a = NA,
  f = NA,
  l = NA,
  follow_up = "until_end"
)
}
\arguments{
\item{Ne}{A vector. Number of events}

\item{r}{A vector. Ratio of sample sizes of treatment group to control group. Default value is 1.}

\item{lambda0}{A vector. Hazard of control group.}

\item{lambda1}{A vector. Hazard of treatment group.}

\item{dropoutRate}{A vector. Dropout rate within the time interval specified by \code{dropoutTime} parameter.}

\item{dropoutTime}{A vector. Time interval for dropout rate.}

\item{a}{A vector. Accrual time for trial, which is only used when \code{follow_up = 'until_end'}.}

\item{f}{A vector. Follow up time for trial, which is only used when \code{follow_up = 'until_end'}.}

\item{l}{A vector. Fixed follow up period for each subject, which is only used when \code{follow_up = 'fixed_period'}.}

\item{follow_up}{A vector. If \code{follow_up = 'until_end'}, subjects will be followed up until the end of trial. If \code{follow_up = 'fixed_period'}, each subject will be followed up a fixed period.}
}
\value{
A data frame containing input parameters and returned event rate and required sample size.
\itemize{
  \item{\code{eventRate0 }}{Event rate for control group.}
  \item{\code{eventRate1 }}{Event rate for treatment group.}
  \item{\code{eventRate }}{Event rate for trial.}
  \item{\code{N0 }}{Required sample size for control group.}
  \item{\code{N1 }}{Required sample size for treatment group.}
  \item{\code{N }}{Required sample size for trial.}
  }
}
\description{
Calculate required sample size according to required number of events, given hazard of event and follow-up parameters in survival analysis. Assuming uniform enrollment of subject and the event time and dropout time follow an exponential distribution.
}
\examples{
# Median survival time in control group is 20 months, HR = 0.75, and annual dropout rate is 5\%.
# Accrual time is 18 months, and follow-up time is 18 months.
# Each subject is followed up until the end of trial.
Ne_to_N(
  Ne = 100, r = 1, lambda0 = log(2) / 20, lambda1 = log(2) / 20 * 0.75,
  dropoutRate = 0.05, dropoutTime = 12,
  a = 18, f = 18, follow_up = "until_end"
)

# Each subject is followed for 18 months after enrollment.
Ne_to_N(
  Ne = 100, r = 1, lambda0 = log(2) / 20, lambda1 = log(2) / 20 * 0.75,
  dropoutRate = 0.05, dropoutTime = 12,
  l = 18, follow_up = "fixed_period"
)
}
\references{
1. Quan H, Zhao PL, Zhang J, Roessner M, Aizawa K. Sample size considerations for Japanese patients in a multi-regional trial based on MHLW guidance. Pharm Stat. 2010;9(2):100-112. doi:10.1002/pst.380
}
