% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regFrac.R
\name{regFrac}
\alias{regFrac}
\title{Regional fraction for one MRCT}
\usage{
regFrac(alpha, power, pi = 0.5, CP)
}
\arguments{
\item{alpha}{The type I error.}

\item{power}{Power.}

\item{pi}{The threshold ratio in Japan's criterion I (conditional version). Defaults to 0.5.}

\item{CP}{the (conditional) consistency probability.}
}
\value{
rF The regional fraction, a scalar.
}
\description{
Calculate the optimal regional fraction given the (conditional) consistency probability for one MRCT.
}
\details{
Given the (conditional) consistency probability, there is a minimum regional
fraction \code{rF}. To calculate the minimum \code{rF}, \code{regFrac} utilizes two core computational components:
\itemize{
\item The \code{\link{conProb}} function to compute the (conditional) consistency probability
\item The \code{\link[stats]{uniroot}} function from the \pkg{stats} package for numerical root-finding
}
The solution is obtained by solving the following equation numerically:
\deqn{
\begin{array}{c}
  \code{conProb(rF)} - \code{CP} = 0 \\
  \text{where } \code{rF} \in (0,1).
\end{array}
}
}
\examples{

regFrac(alpha=0.025,power=0.8,pi=0.5,CP=0.8)

}
