% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regFrac2MRCTs.R
\name{regFrac2MRCTs}
\alias{regFrac2MRCTs}
\title{Regional fraction for two MRCTs in general}
\usage{
regFrac2MRCTs(
  alpha,
  power1,
  power2,
  pi,
  CP,
  d1,
  d2,
  sigmaTrt1,
  sigmaCtrl1,
  sigmaTrt2,
  sigmaCtrl2,
  ratio1,
  ratio2
)
}
\arguments{
\item{alpha}{The type I error.}

\item{power1}{Power for MRCT 1.}

\item{power2}{Power for MRCT 2.}

\item{pi}{The threshold ratio in Japan's criterion I (conditional version).}

\item{CP}{The consistency probability.}

\item{d1}{The true mean of difference of response for MRCT 1.}

\item{d2}{The true mean of difference of response for MRCT 2.}

\item{sigmaTrt1}{The standard deviation of response in the treatment group for MRCT 1.}

\item{sigmaCtrl1}{The standard deviation of response in the control group for MRCT 1.}

\item{sigmaTrt2}{The standard deviation of response in the treatment group for MRCT 2.}

\item{sigmaCtrl2}{The standard deviation of response in the control group for MRCT 2.}

\item{ratio1}{The randomization ratio between the treatment group and control group for MRCT 1.}

\item{ratio2}{The randomization ratio between the treatment group and control group for MRCT 2.}
}
\value{
A list containing the following two components:
\describe{
\item{\code{rF1}}{The optimal regional fraction for MRCT 1.}
\item{\code{rF2}}{The optimal regional fraction for MRCT 2.}
}
}
\description{
Calculate the optimal regional fractions given the (conditional) consistency probability for two MRCT in general.
}
\details{
Given the (conditional) consistency probability, there is an optimal pair of regional
fractions \code{rF1} and \code{rF2} that minimized the combined region sample size,
i.e., \eqn{f_{k}^{(1)}N^{(1)} + f_{k}^{(2)}N^{(2)}}. Theoretically,
the ratio between such \code{rF1} and \code{rF2} is fixed. Hence,
\code{regFrac2MRCTs} utilizes two core computational components:
\itemize{
\item The \code{\link{conProb2MRCTs}} function to compute the (conditional) consistency probability for two MRCTs
\item The \code{\link[stats]{uniroot}} function from the \pkg{stats} package for numerical root-finding
}

The solution is obtained by solving the following equation numerically:
\deqn{
\begin{array}{c}
  \code{conProb2MRCTs(rF2*k,rF2)} - \code{CP} = 0 \\
  \text{where } \code{rF2} \in (0,\min(1/k,1)),
\end{array}
}
where \eqn{k} is the above fixed ratio between \code{rF1} and \code{rF2}.
}
\examples{
\dontrun{
regFrac2MRCTs(0.025,0.8,0.8,0.5,0.8,1,1,4,4,4,4,1,1)
}

}
