% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2MRCTs.R
\name{table2MRCTs}
\alias{table2MRCTs}
\title{Tables for two MRCTs paper}
\usage{
table2MRCTs(numSimu, choice, seed = 1)
}
\arguments{
\item{numSimu}{The number of simulation times. Defaults to 100,000}

\item{choice}{Table number.}

\item{seed}{Random seed. Defaults to 1.}
}
\value{
Tables shown in the reference
}
\description{
Tables for two MRCTs paper
}
\examples{
table2MRCTs(100,1,1)

}
