% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{combine}
\alias{combine}
\title{Combine vectors or matrices}
\usage{
combine(A, B)
}
\arguments{
\item{A}{A vector or matrix}

\item{B}{A vector or matrix}
}
\value{
A vector or matrix
}
\description{
Combine two vectors to one vector or two matrices to one matrix
}
\examples{
combine(c(1, 2, 3), c(4, 5, 6))

combine(
  matrix(c(1, 2, 3, 4), nrow = 2, byrow = TRUE),
  matrix(c(5, 6, 7, 8), nrow = 2, byrow = TRUE)
)
}
