% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conProb.R
\name{conProb}
\alias{conProb}
\title{Consistency probability for one MRCT}
\usage{
conProb(alpha, power, pi = 0.5, rF)
}
\arguments{
\item{alpha}{The Type I error.}

\item{power}{Power.}

\item{pi}{The threshold ratio in Japan's criterion I (conditional version). Defaults to 0.5.}

\item{rF}{The regional fraction.}
}
\value{
CP The consistency probability, a scalar.
}
\description{
Calculate the (conditional) consistency probability for one MRCT.
}
\details{
The (conditional) consistency probability, \eqn{\mathrm{Pr}\left(D_{k}\ge \pi D\ |\ T>z_{1-\alpha}\right)},
equals
\deqn{
 \frac{1}{1-\beta}\int_{-z_{1-\beta}}^{\infty} \Phi\left(\frac{(1-\pi)(u + z_{1-\alpha}+z_{1-\beta})}{\sqrt{f_{k}^{-1}-1}}\right)\phi(u)du.
}
Since there is no closed form of above equation, \code{conProb} utilizes the \code{\link[stats]{integrate}} function for numerical integration.
}
\examples{
conProb(alpha=0.025,power=0.8,pi=0.5,rF=0.23)

}
