% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conProb2.R
\name{conProb2}
\alias{conProb2}
\title{Consistency probability for two MRCTs}
\usage{
conProb2(
  alpha,
  power1,
  power2 = power1,
  pi = 0.5,
  rF1,
  rF2 = rF1,
  d1,
  d2 = d1,
  sigmaTrt1,
  sigmaCtrl1 = sigmaTrt1,
  sigmaTrt2 = sigmaTrt1,
  sigmaCtrl2 = sigmaTrt2,
  randRatio1 = 1,
  randRatio2 = randRatio1
)
}
\arguments{
\item{alpha}{The type I error.}

\item{power1}{Power for MRCT 1.}

\item{power2}{Power for MRCT 2. Defaults to \code{power1}.}

\item{pi}{The threshold ratio in Japan's criterion I (conditional version). Defaults to 0.5.}

\item{rF1}{The regional fraction for MRCT 1.}

\item{rF2}{The regional fraction for MRCT 2. Defaults to \code{rF1}.}

\item{d1}{The true mean of difference of response for MRCT 1.}

\item{d2}{The true mean of difference of response for MRCT 2. Defaults to \code{d1}.}

\item{sigmaTrt1}{The standard deviation of response in the treatment group for MRCT 1.}

\item{sigmaCtrl1}{The standard deviation of response in the control group for MRCT 1. Defaults to \code{sigmaTrt1}.}

\item{sigmaTrt2}{The standard deviation of response in the treatment group for MRCT 2. Defaults to \code{sigmaTrt1}.}

\item{sigmaCtrl2}{The standard deviation of response in the control group for MRCT 2. Defaults to \code{sigmaTrt2}.}

\item{randRatio1}{The randomization ratio between the treatment group and control group for MRCT 1.}

\item{randRatio2}{The randomization ratio between the treatment group and control group for MRCT 2. Defaults to \code{randRatio1}.}
}
\value{
A list containing the following two components:
\describe{
\item{\code{CP}}{The consistency probability, a scalar.}
\item{\code{N1}}{The overall sample size for MRCT 1.}
\item{\code{N2}}{The overall sample size for MRCT 2.}
}
}
\description{
Calculate the (conditional) consistency probability for two MRCTs.
}
\details{
The extended (conditional) consistency probability, \eqn{\mathrm{Pr}\left(D_{k,\textrm{pool}}\ge \pi D_\textrm{pool}\ |\ T^{(1)}>z_{1-\alpha},T^{(2)}>z_{1-\alpha}\right)},
equals
\deqn{
 \begin{aligned}
 &\frac{1}{(1-\beta_{1})(1-\beta_{2})}\int_{-z_{1-\beta_{1}}}^{\infty}\int_{-z_{1-\beta_{2}}}^{\infty} \\
 &\quad
 \Phi\left(\frac{(1-\pi)\left\{w^{(1)}\sigma^{(1)}_{d}u+w^{(2)}\sigma^{(2)}_{d}v + w^{(1)}d^{(1)} + w^{(2)}d^{(2)}\right\}}{\sqrt{\left\{\left(f_{k}^{(1)}\right)^{-1}-1\right\} \left(w^{(1)}\sigma^{(1)}_{d}\right)^{2} + \left\{\left(f_{k}^{(2)}\right)^{-1}-1\right\} \left(w^{(2)}\sigma^{(2)}_{d}\right)^{2}}} \right)\phi(u)\phi(v)dudv,
 \end{aligned}
}
where \eqn{w^{(s)}=N^{(s)}/(N^{(1)}+N^{(2)})} and
\deqn{
 \sigma_d^{2}=\var(D)=\frac{(r+1)\left\{\sigma^{2(\textrm{t})}+r\sigma^{2(\textrm{c})}\right\}}{rN}.
}

Since there is no closed forms of above equations, \code{conProb2} utilizes
the \code{\link[cubature]{adaptIntegrate}} function for numerical integration.

Additionally, if we assume homogeneous variances, equal treatment effects
and equal randomization ratios across two studies, i.e., \eqn{\sigma^{2(\textrm{t},1)}=\sigma^{2(\textrm{t},2)}},
\eqn{\sigma^{2(\textrm{c},1)}=\sigma^{2(\textrm{c},2)}}, \eqn{d^{(1)}=d^{(2)}}
and \eqn{r^{(1)}=r^{(2)}}, then the above equation reduces to
\deqn{
 \begin{aligned}
 &\frac{1}{(1-\beta_{1})(1-\beta_{2})}\int_{-(z_{1-\beta_{1}}+z_{1-\beta_{2}})/\sqrt{2}}^{\infty}\left[ \Phi\left(\frac{(1-\pi)\left\{\sqrt{2}u + (2z_{1-\alpha}+z_{1-\beta_{1}}+z_{1-\beta_{2}})\right\}}{\sqrt{\left(f_{k}^{(1)}\right)^{-1}+\left(f_{k}^{(2)}\right)^{-1}-2}}\right) \right. \\
 &\qquad\qquad\qquad\qquad \left.\left\{ \Phi\left(u+\sqrt{2}z_{1-\beta_{1}}\right)+\Phi\left(u+\sqrt{2}z_{1-\beta_{2}}\right)-1\right\} \right] \phi(u)du.
 \end{aligned}
}

Since now the above equation includes none of \code{d1}=\code{d2},
\code{sigmaTrt1}=\code{sigmaTrt2}, \code{sigmaCtrl1}=\code{sigmaCtrl2}
\code{randRatio1} = \code{randRatio2}, if only the consistency probability
is considered, then the values of \code{d1} and \code{sigmaTrt1} could be arbitrary.

The overall sample size is calculated in the same way as \code{\link{conProb}}.
}
\examples{

conProb2(alpha = 0.025, power1 = 0.8, rF1 = 0.127, d1 = 1, sigmaTrt1 = 4)

}
