% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/33 getN_Con_Equi.R
\name{getN_Con_Equi}
\alias{getN_Con_Equi}
\title{Title}
\usage{
getN_Con_Equi(delta, sigma, cut, alpha, beta, N, r, maxN = 1e+06)
}
\arguments{
\item{maxN}{}
}
\description{
Title
}
\examples{
(v <- getN_Con_Equi(
  delta = 0, sigma = 2, cut = 1, alpha = 0.025,
  beta = 0.2, N = NA, r = 1
))
getN_Con_Equi(
  delta = 0, sigma = 2, cut = 1, alpha = 0.025,
  beta = NA, N = v$N, r = 1
)
}
