% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3 getPwr_Con_Equi_JM1.R
\name{getPwr_Con_Equi_JM1}
\alias{getPwr_Con_Equi_JM1}
\title{Title}
\usage{
getPwr_Con_Equi_JM1(
  delta_j,
  delta_nj,
  sigma,
  f,
  pi,
  cut,
  alpha,
  N,
  r,
  sim = FALSE,
  nsim = 1000,
  seed = 0
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
\examples{
getPwr_Con_Equi_JM1(
  delta_j = -0.2, delta_nj = -0.1, sigma = 1,
  f = seq(0.1, 0.9, 0.1), pi = 0.5, cut = 0.4, alpha = 0.025,
  N = 400, r = 1, sim = FALSE
)
}
