% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1 getPwr_Con_Super_JM1.R
\name{getPwr_Con_Super_JM1}
\alias{getPwr_Con_Super_JM1}
\title{Title}
\usage{
getPwr_Con_Super_JM1(
  delta_j,
  delta_nj,
  sigma,
  f,
  pi = 0.5,
  alpha = 0.025,
  beta = NA,
  N,
  r = 1,
  sim = FALSE,
  nsim = 1000,
  seed = 0,
  numcore = 2
)
}
\arguments{
\item{numcore}{}
}
\description{
Title
}
\examples{
getPwr_Con_Super_JM1(
  delta_j = 0.5, delta_nj = 0.7, sigma = 1,
  f = seq(0.1, 0.9, 0.1),
  pi = 0.5, alpha = 0.025, beta = NA, N = 100, r = 1, sim = FALSE
)
getPwr_Con_Super_JM1(
  delta_j = 0.5, delta_nj = 0.7, sigma = 1,
  f = seq(0.1, 0.9, 0.1),
  pi = 0.5, alpha = 0.025, beta = 0.2, N = NA, r = 1, sim = FALSE
)
}
