% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4 getPwr_Con_Super_JM2.R
\name{getPwr_Con_Super_JM2}
\alias{getPwr_Con_Super_JM2}
\title{Title}
\usage{
getPwr_Con_Super_JM2(
  delta_i,
  sigma,
  fi,
  alpha = 0.025,
  beta = NA,
  N,
  r = 1,
  sim = FALSE,
  nsim = 1000,
  seed = 0
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
