% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13 getPwr_Surv_Super_JM1.R
\name{getPwr_Surv_Super_JM1}
\alias{getPwr_Surv_Super_JM1}
\title{Title}
\usage{
getPwr_Surv_Super_JM1(
  delta_j,
  delta_nj,
  f,
  pi,
  alpha,
  N,
  r,
  criterion,
  direct = -1,
  lambda0_j = 1,
  lambda0_nj = 1,
  sim = FALSE,
  nsim = 1000,
  seed = 0
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
\examples{
getPwr_Surv_Super_JM1(
  delta_j = log(0.8), delta_nj = log(0.6),
  f = seq(0.1, 0.9, 0.1),
  pi = 0.5, alpha = 0.025, N = 200, r = 1,
  criterion = 1, direct = -1, sim = FALSE
)
}
