% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regFrac.R
\name{regFrac}
\alias{regFrac}
\title{Regional fraction for MRCT via criterion Method I of MHLW (2007) (conditional version) under fixed effects model}
\usage{
regFrac(
  alpha,
  power,
  pi = 0.5,
  CP = 0.8,
  d,
  sigmaTrt,
  sigmaCtrl = sigmaTrt,
  randRatio = 1
)
}
\arguments{
\item{alpha}{The Type I error.}

\item{power}{Power.}

\item{pi}{The threshold ratio in criterion Method I of MHLW (2007) (conditional version). Defaults to 0.5.}

\item{CP}{the consistency probability. Defaults to 80\%.}

\item{d}{The true mean of difference of response.}

\item{sigmaTrt}{The standard deviation of response in the treatment group.}

\item{sigmaCtrl}{The standard deviation of response in the control group. Defaults to \code{sigmaTrt}.}

\item{randRatio}{The randomization ratio between the treatment group and control group. Defaults to 1.}
}
\value{
A list containing the following two components:
\describe{
\item{\code{rF}}{The regional fraction, a scalar.}
\item{\code{N}}{The overall sample size.}
}
}
\description{
Calculate the minimal regional fraction given the consistency probability for MRCT via criterion Method I of MHLW (2007) (conditional version) under fixed effects model.
}
\details{
Given the consistency probability, there is a minimal regional
fraction \code{rF}. To calculate the minimal \code{rF}, \code{regFrac} utilizes two core computational components:
\itemize{
\item The \code{\link{conProb}} function to compute the (conditional) consistency probability
\item The \code{\link[stats]{uniroot}} function from the \pkg{stats} package for numerical root-finding
}

The solution is obtained by solving the following equation numerically:
\deqn{
\begin{array}{c}
  \code{conProb(rF)} - \code{CP} = 0 \\
  \text{where } \code{rF} \in (0,1).
\end{array}
}

The overall sample size is obtain by \code{\link{conProb}} simultaneously.
}
\examples{
conProb(alpha = 0.05, power = 0.8, rF = 0.271, d = 1, sigmaTrt = 4)

regFrac(alpha = 0.05, power = 0.8, d = 1, sigmaTrt = 4)
  
}
