% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regFrac2.R
\name{regFrac2}
\alias{regFrac2}
\title{Regional fraction for two MRCTs via the extended criterion Method I of MHLW (2007) (conditional version) under fixed effects model}
\usage{
regFrac2(
  alpha,
  power1,
  power2 = power1,
  pi = 0.5,
  CP = 0.8,
  d1,
  d2 = d1,
  sigmaTrt1,
  sigmaCtrl1 = sigmaTrt1,
  sigmaTrt2 = sigmaTrt1,
  sigmaCtrl2 = sigmaTrt2,
  randRatio1 = 1,
  randRatio2 = randRatio1
)
}
\arguments{
\item{alpha}{The Type I error.}

\item{power1}{Power for MRCT 1.}

\item{power2}{Power for MRCT 2. Defaults to \code{power1}.}

\item{pi}{The threshold ratio in the extended criterion Method I of MHLW (2007) (conditional version). Defaults to 0.5.}

\item{CP}{The consistency probability. Defaults to 80\%.}

\item{d1}{The true mean of difference of response for MRCT 1.}

\item{d2}{The true mean of difference of response for MRCT 2. Defaults to \code{d1}.}

\item{sigmaTrt1}{The standard deviation of response in the treatment group for MRCT 1.}

\item{sigmaCtrl1}{The standard deviation of response in the control group for MRCT 1. Defaults to \code{sigmaTrt1}.}

\item{sigmaTrt2}{The standard deviation of response in the treatment group for MRCT 2. Defaults to \code{sigmaTrt1}.}

\item{sigmaCtrl2}{The standard deviation of response in the control group for MRCT 2. Defaults to \code{sigmaTrt2}.}

\item{randRatio1}{The randomization ratio between the treatment group and control group for MRCT 1.}

\item{randRatio2}{The randomization ratio between the treatment group and control group for MRCT 2. Defaults to \code{randRatio1}.}
}
\value{
A list containing the following two components:
\describe{
\item{\code{rF1}}{The regional fraction for MRCT 1, a scalar.}
\item{\code{rF2}}{The regional fraction for MRCT 2, a scalar.}
\item{\code{N1}}{The overall sample size for MRCT 1, an integer.}
\item{\code{N2}}{The overall sample size for MRCT 2, an integer.}
}
}
\description{
Calculate the optimal regional fractions given the consistency probability for two MRCTs via the extended criterion Method I of MHLW (2007)  (conditional version) under fixed effects model
}
\details{
Given the consistency probability, there is an optimal pair of regional
fractions \code{rF1} and \code{rF2} that minimized the combined region sample size,
i.e., \eqn{f_{k}^{(1)}N^{(1)} + f_{k}^{(2)}N^{(2)}}. Theoretically,
the ratio between such \code{rF1} and \code{rF2} is fixed. Hence,
\code{regFrac2} utilizes two core computational components:
\itemize{
\item The \code{\link{conProb2}} function to compute the (conditional) consistency probability for two MRCTs
\item The \code{\link[stats]{uniroot}} function from the \pkg{stats} package for numerical root-finding
}

The solution is obtained by solving the following equation numerically:
\deqn{
\begin{array}{c}
  \code{conProb2(rF2*k,rF2)} - \code{CP} = 0 \\
  \text{where } \code{rF2} \in (0,\min(1/k,1)),
\end{array}
}
where \eqn{k} is the above fixed ratio between \code{rF1} and \code{rF2}.

The overall sample size is obtain by \code{\link{conProb2}} simultaneously.
}
\examples{
### Remark 7
conProb2(alpha = 0.05, power1 = 0.8, power2 = 0.9, rF1 = 0.141, d1 = 1, sigmaTrt1 = 4) 

regFrac2(alpha = 0.05, power1 = 0.8, power2 = 0.9, d1 = 1, sigmaTrt1 = 4)

conProb2(alpha = 0.05, power1 = 0.8, power2 = 0.9, rF1=0.1407622, d1 = 1, sigmaTrt1 = 4)

}
